hermes::cli

Module playground

Simplified runner for Hermes WASM components. This runs each wasm file as a separate anonymous application.

Structs

Playground

Hermes application playground

Functions

collect_apps 🔒

Collects .wasm files in the current directory or sub-directories of the current directory. Then creates one-module applications out of each of them.

collect_modules 🔒

Collects .wasm files in the current directory or sub-directories of the current directory. Return a String module name along with each compiled Module.

create_one_module_app 🔒

Create one-module application with temp directory VFS.

create_temp_dir_child 🔒

Create a temp subdirectory.

init_ipfs 🔒

Initialize the IPFS node