Module hermes::event::queue

Hermes event queue implementation.

Structs

AlreadyInitializedError 🔒

Failed when event queue already been initialized.

CannotAddEventError 🔒

Failed to add event into the event queue. Event queue is closed.

EventLoopPanicsError 🔒

Event loop has crashed unexpectedly.

HermesEventLoopHandler 🔒

Hermes event queue execution loop thread handler

HermesEventQueue 🔒

Hermes event queue. It is a singleton struct.

NotInitializedError 🔒

Failed when event queue not been initialized.

Statics

EVENT_QUEUE_INSTANCE 🔒

Singleton instance of the Hermes event queue.

Functions

event_dispatch 🔒

Execute a hermes event on the provided module and all necessary info.

event_execution_loop 🔒

Executes Hermes events from the provided receiver .

get_execution_context 🔒

Get execution context

init 🔒

Creates a new instance of the HermesEventQueue. Runs an event loop thread.

send 🔒

Add event into the event queue

targeted_event_execution 🔒

Executes provided Hermes event filtering by target app and target module.

Type Aliases

ExecutionContext 🔒

Hermes event execution context