Module hermes::packaging

Hermes packaging.

Modules

app 🔒

Hermes application package.

hash 🔒

Blake2b-256 hash implementation.

metadata 🔒

A generalized metadata object which is used an specified by the Hermes application package or Hermes WASM module package.

package 🔒

Implementation of the generalized Hermes package struct as a HDF5 package.

resources 🔒

Resources module functionality.

schema_validation 🔒

JSON schema validation module.

sign 🔒

Hermes signing procedure implementation. Follows documented spec https://input-output-hk.github.io/hermes/architecture/08_concepts/hermes_signing_procedure.

wasm_module 🔒

Hermes WASM module package.

Structs

FileError 🔒

File open and read error.

MissingPackageFileError 🔒

Missing package file error.