hermes::packaging

Module app

Hermes application package.

Modules

app_builder 🔒

Application builder from the application package.

author_payload 🔒

A signature payload object for author.cose Hermes application package. Defined at https://input-output-hk.github.io/hermes/architecture/08_concepts/hermes_packaging_requirements/app_signatures/#wasm-component-module-signatures.

manifest 🔒

Hermes application package manifest.json struct.

module_info 🔒

An application’s module info object

Structs

ApplicationPackage 🔒

Hermes application package.

Functions

is_svg_file 🔒

Validate svg file.

validate_and_write_icon 🔒

Validate icon.svg file and write it to the package to the provided dir path.

validate_and_write_metadata 🔒

Validate metadata.json file and write it to the package to the provided dir path. Also updates Metadata object by setting build_date and name properties.

validate_and_write_module 🔒

Validate WASM module package and write it to the package to the provided dir path.

write_share_dir 🔒

Write share dir to the package to the provided dir path.

write_www_dir 🔒

Write www dir to the package to the provided dir path to the provided dir path.