hermes::packaging::app

Struct ApplicationPackage

pub(crate) struct ApplicationPackage(Package);

Hermes application package.

Tuple Fields
0: Package
Implementations

impl ApplicationPackage

const AUTHOR_COSE_FILE: &'static str = "author.cose"

Application package signature file path.

const FILE_EXTENSION: &'static str = "happ"

Application package file extension.

const ICON_FILE: &'static str = "icon.svg"

Application package icon file path.

const LIB_DIR: &'static str = "lib"

Application package ‘lib’ directory path.

const METADATA_FILE: &'static str = "metadata.json"

Application package metadata file path.

const MODULE_CONFIG_FILE: &'static str = "config.json"

Application package overridden module’s config file name.

const MODULE_SHARE_DIR: &'static str = "share"

Application package overridden module’s ‘share’ dir name.

const SRV_DIR: &'static str = "srv"

Application package srv directory name.

const SRV_SHARE_DIR: &'static str = "srv/share"

Application package srv/share directory path.

const SRV_WWW_DIR: &'static str = "srv/www"

Application package srv/www directory path.

const USR_DIR: &'static str = "usr"

Application package ‘usr’ directory path.

const USR_LIB_DIR: &'static str = "usr/lib"

Application package ‘usr/lib’ directory path.

pub(crate) fn build_from_manifest<P: AsRef<Path>>( manifest: &Manifest, output_path: P, package_name: Option<&str>, build_date: DateTime<Utc>, ) -> Result<Self>

Create a new Hermes application package package from a manifest file.

pub(crate) fn from_file<P: AsRef<Path>>(path: P) -> Result<Self>

Open an existing application package.

pub(crate) fn validate(&self, untrusted: bool) -> Result<()>

Validate package with its signature and other contents. If untrusted flag is true the signature will not be verified.

fn verify_author_sign(&self) -> Result<()>

Verify author package signature.

pub(crate) fn author_sign( &self, private_key: &PrivateKey, certificate: &Certificate, ) -> Result<()>

Sign the package as an author and store signature inside it. If signature already exists it will be extended with a new signature.

fn get_author_signature_payload(&self) -> Result<SignaturePayload>

Build and return author_payload::SignaturePayload.

pub(crate) fn get_app_name(&self) -> Result<String>

Get application package name.

fn get_icon_file(&self) -> Result<File>

Get icon File object from package.

fn get_metadata_file(&self) -> Result<File>

Get metadata File object from package.

pub(crate) fn get_metadata(&self) -> Result<Metadata<Self>>

Get Metadata object from package.

pub(crate) fn get_author_signature( &self, ) -> Result<Option<Signature<SignaturePayload>>>

Get author Signature object from package.

pub(crate) fn get_modules(&self) -> Result<Vec<AppModuleInfo>>

Get Vec<WasmModulePackage> from package.

fn get_www_dir(&self) -> Option<Dir>

Get www dir from package if present.

fn get_share_dir(&self) -> Option<Dir>

Get share dir from package if present.

fn validate_and_write_from_manifest( manifest: &Manifest, package: &Package, build_date: DateTime<Utc>, package_name: &str, errors: &mut Errors, )

Validate and write all content of the Manifest to the provided package.