hermes::packaging

Module module

Hermes WASM module package.

Modules

author_payload 🔒

A signature payload object for author.cose Hermes WASM module package. Defined at https://input-output-hk.github.io/hermes/architecture/08_concepts/hermes_packaging_requirements/wasm_modules/#wasm-component-module-signatures.

config 🔒

WASM module package config JSON

config_info 🔒

Hermes WASM module’s config info object.

manifest 🔒

WASM module package manifest.json struct.

settings 🔒

WASM module package settings JSON

Structs

ModulePackage 🔒

Hermes WASM module package.

Functions

validate_and_write_component 🔒

Validate WASM component file and write it to the package to the provided dir path.

validate_and_write_config 🔒

Validate config schema and config file and write them to the package.

validate_and_write_config_file 🔒

Validate config file and write it to the package.

validate_and_write_config_schema 🔒

Validate config schema and write it to the package to the provided dir path.

validate_and_write_metadata 🔒

Validate metadata.json file and write it to the package to the provided dir path. Also updates Metadata object by setting build_date and name properties.

validate_and_write_settings_schema 🔒

Validate settings schema file and it to the package.

write_share_dir 🔒

Write share dir to the package.