hermes::packaging::module

Struct ModulePackage

pub(crate) struct ModulePackage(Package);

Hermes WASM module package.

Tuple Fields
0: Package
Implementations

impl ModulePackage

const AUTHOR_COSE_FILE: &'static str = "author.cose"

Module package signature file path.

const COMPONENT_FILE: &'static str = "module.wasm"

Module package WASM component file path.

const CONFIG_FILE: &'static str = "config.json"

Module package config file path.

const CONFIG_SCHEMA_FILE: &'static str = "config.schema.json"

Module package config schema file path.

pub(crate) const FILE_EXTENSION: &'static str = "hmod"

Module package file extension.

const METADATA_FILE: &'static str = "metadata.json"

Module package metadata file path.

const SETTINGS_SCHEMA_FILE: &'static str = "settings.schema.json"

Module package settings schema file path.

const SHARE_DIR: &'static str = "share"

Module package share directory path.

pub(crate) fn build_from_manifest<P: AsRef<Path>>( manifest: &Manifest, output_path: P, package_name: Option<&str>, build_date: DateTime<Utc>, ) -> Result<Self>

Create a new module package from a manifest file.

pub(crate) fn from_file<P: AsRef<Path>>(path: P) -> Result<Self>

Open an existing WASM module package.

pub(crate) fn from_package(package: Package) -> Self

Create ModulePackage from a Package.

pub(crate) fn validate(&self, untrusted: bool) -> Result<()>

Validate package with its signature and other contents. If untrusted flag is true the signature will not be verified.

fn verify_sign(&self) -> Result<()>

Verify package signature.

pub(crate) fn sign( &self, private_key: &PrivateKey, certificate: &Certificate, ) -> Result<()>

Sign the package and store signature inside it. If signature already exists it will be extended with a new signature.

fn get_signature_payload(&self) -> Result<SignaturePayload>

Build and return SignaturePayload.

pub(super) fn get_metadata_file(&self) -> Result<File>

Get metadata File object from package.

pub(crate) fn get_metadata(&self) -> Result<Metadata<Self>>

Get Metadata object from package.

pub(super) fn get_component_file(&self) -> Result<File>

Get component File object from package.

pub(crate) fn get_component(&self, app_name: &ApplicationName) -> Result<Module>

Get wasm::module::Module object from package.

pub(crate) fn get_signature( &self, ) -> Result<Option<Signature<SignaturePayload>>>

Get Signature object from package.

pub(super) fn get_config_schema_file(&self) -> Option<File>

Get config schema File object from package.

pub(crate) fn get_config_schema(&self) -> Result<Option<ConfigSchema>>

Get ConfigSchema object from package.

pub(super) fn get_config_file(&self) -> Option<File>

Get config File object from package.

pub(crate) fn get_config_info(&self) -> Result<Option<ConfigInfo>>

Get Config and ConfigSchema objects from package if present.

pub(super) fn get_settings_schema_file(&self) -> Option<File>

Get settings schema File object from package if present.

pub(crate) fn get_settings_schema(&self) -> Result<Option<SettingsSchema>>

Get SettingsSchema object from package if present.

pub(super) fn get_share_dir(&self) -> Option<Dir>

Get share dir from package if present.

pub(crate) fn copy_to_dir(&self, dir: &Dir, path: &Path) -> Result<()>

Copy all content of the ModulePackage to the provided Dir.

fn validate_and_write_from_manifest( manifest: &Manifest, package: &Package, build_date: DateTime<Utc>, package_name: &str, errors: &mut Errors, )

Validate and write all content of the Manifest to the provided package.