Module hermes::packaging::package::compression

Enabling blosc compression for the hdf5 package.

Constants

COMPRESSION_ALGORITHM 🔒

Compression algorithm.

COMPRESSION_LEVEL 🔒

Compression level. zstd compressor of 9 to get a good balance of compression ration and decompression speed.

MIN_BLOSC_THREADS 🔒

Minimum blosc threads. 8 threads being a minimum likely standard across most computers.

MIN_CHUNK_SIZE 🔒

Minimum chunk size in kb, 8mb. 1mb per compressor/decompressor thread

Statics

BLOSC_THREADS_INIT 🔒

Statically initialize blosc on the first call only once.

Functions

blosc_init 🔒

Initialize blosc.

enable_compression 🔒

Enable blosc compression.