Module hermes::packaging::wasm_module

Hermes WASM module package.

Modules

config 🔒

WASM package config JSON

manifest 🔒

WASM module package manifest.json struct.

settings 🔒

WASM package settings JSON

signature_payload 🔒

A signature payload object for WASM module package. Defined at https://input-output-hk.github.io/hermes/architecture/08_concepts/hermes_packaging_requirements/wasm_modules/#wasm-component-module-signatures.

Structs

WasmModulePackage 🔒

Hermes WASM module package.

Functions

validate_and_write_component 🔒

Validate WASM component file and write it to the package.

validate_and_write_config 🔒

Validate config file and config schema and write them to the package.

validate_and_write_metadata 🔒

Validate metadata.json file and write it to the package. Also updates Metadata object by setting build_date and name properties.

validate_and_write_settings 🔒

Validate settings schema file and it to the package.

write_share_dir 🔒

Write share dir to the package.