hermes

Module reactor

Hermes Reactor implementation.

Structs

AlreadyInitializedError 🔒

Failed when reactor already been initialized.

NotInitializedError 🔒

Failed when event queue not been initialized.

Reactor 🔒

Hermes Reactor struct. This object orchestrates all Hermes apps within all core parts of the Hermes.

Statics

REACTOR_STATE 🔒

Global Hermes reactor state

Functions

get_all_app_names 🔒

Get all available Hermes application names from the Hermes Reactor.

get_app 🔒

Get Hermes application from the Hermes Reactor.

init 🔒

Initialize Hermes Reactor. Setup and runs all necessary services.

load_app 🔒

Load Hermes application into the Hermes Reactor.