Trait hermes::runtime::extensions::hermes::hash::api::Host

pub trait Host {
    // Required methods
    fn blake2s(
        &mut self,
        buf: Vec<u8>,
        outlen: Option<u8>,
        key: Option<Vec<u8>>
    ) -> Result<Result<Vec<u8>, Errno>>;
    fn blake2b(
        &mut self,
        buf: Vec<u8>,
        outlen: Option<u8>,
        key: Option<Vec<u8>>
    ) -> Result<Result<Vec<u8>, Errno>>;
    fn blake3(
        &mut self,
        buf: Vec<u8>,
        outlen: Option<u8>,
        key: Option<Vec<u8>>
    ) -> Result<Result<Vec<u8>, Errno>>;
}

Required Methods

fn blake2s( &mut self, buf: Vec<u8>, outlen: Option<u8>, key: Option<Vec<u8>> ) -> Result<Result<Vec<u8>, Errno>>

BLAKE2s Hash Function

Hash a binary buffer with BLAKE2s.

Parameters

buf: The binary data buffer to hash.

outlen: Optional. The size of the digest. If the outlen is not defined, it defaults to 32.

key: Optional. The key to use with the hash. If not defined, the hash is not keyed. Should not be > 32 bytes.

Returns

Either a buffer the size requested, with the hash.

Or an error:

key_too_big if key is specified and is > 32 bytes.

hash_too_big if outlen is specified and is > 32 bytes.

Note:

key length is checked before outlen so if both sizes are invalid, only key_too_big will be returned.

fn blake2b( &mut self, buf: Vec<u8>, outlen: Option<u8>, key: Option<Vec<u8>> ) -> Result<Result<Vec<u8>, Errno>>

BLAKE2b Hash Function

Hash a binary buffer with BLAKE2b.

Parameters

buf: The binary data buffer to hash.

outlen: Optional. The size of the digest. If the outlen is not defined, it defaults to 64.

key: Optional. The key to use with the hash. If not defined, the hash is not keyed. Should not be > 64 bytes.

Returns

Either a buffer the size requested, with the hash.

Or an error:

key_too_big if key is specified and is > 64 bytes.

hash_too_big if outlen is specified and is > 64 bytes.

Note:

key length is checked before outlen so if both sizes are invalid, only key_too_big will be returned.

fn blake3( &mut self, buf: Vec<u8>, outlen: Option<u8>, key: Option<Vec<u8>> ) -> Result<Result<Vec<u8>, Errno>>

BLAKE3 Hash Function

Hash a binary buffer with BLAKE3.

Parameters

buf: The binary data buffer to hash.

outlen: Optional. The size of the digest. If the outlen is not defined, it defaults to 32.

key: Optional. The key to use with the hash. If not defined, the hash is not keyed. Should not be > 32 bytes.

Returns

Either a buffer the size requested, with the hash.

Or an error:

key_too_big if key is specified and is > 32 bytes.

hash_too_big if outlen is specified and is > 32 bytes.

Note:

key length is checked before outlen so if both sizes are invalid, only key_too_big will be returned.