Struct hermes::runtime::extensions::Hermes

pub struct Hermes {
    interface0: IncomingHandler,
    interface1: EventOnBlock,
    interface2: EventOnTxn,
    interface3: EventOnRollback,
    interface4: Event,
    interface5: Event,
    interface6: Event,
}

Fields
interface0: IncomingHandlerinterface1: EventOnBlockinterface2: EventOnTxninterface3: EventOnRollbackinterface4: Eventinterface5: Eventinterface6: Event
Implementations

impl Hermes

pub fn add_to_linker<T, U>( linker: &mut Linker<T>, get: impl Fn(&mut T) -> &mut U + Send + Sync + Copy + 'static ) -> Result<()>
where U: Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host,

pub fn instantiate<T>( store: impl AsContextMut<Data = T>, component: &Component, linker: &Linker<T> ) -> Result<(Self, Instance)>

Instantiates the provided module using the specified parameters, wrapping up the result in a structure that translates between wasm and the host.

pub fn instantiate_pre<T>( store: impl AsContextMut<Data = T>, instance_pre: &InstancePre<T> ) -> Result<(Self, Instance)>

Instantiates a pre-instantiated module using the specified parameters, wrapping up the result in a structure that translates between wasm and the host.

pub fn new(store: impl AsContextMut, instance: &Instance) -> Result<Self>

Low-level creation wrapper for wrapping up the exports of the instance provided in this structure of wasm exports.

This function will extract exports from the instance defined within store and wrap them all up in the returned structure which can be used to interact with the wasm module.

pub fn wasi_http_incoming_handler(&self) -> &IncomingHandler

pub fn hermes_cardano_event_on_block(&self) -> &EventOnBlock

pub fn hermes_cardano_event_on_txn(&self) -> &EventOnTxn

pub fn hermes_cardano_event_on_rollback(&self) -> &EventOnRollback

pub fn hermes_cron_event(&self) -> &Event

pub fn hermes_init_event(&self) -> &Event

pub fn hermes_kv_store_event(&self) -> &Event