hermes::runtime_extensions

Module bindings

Hermes wasmtime::component::bindgen generated code

Note Inspect the generated code with: cargo expand -p hermes --lib runtime_extensions::bindings or with: earthly +bindings-expand

Modules

exports

hermes

wasi

Structs

Hermes

Auto-generated bindings for an instance a component which implements the world hermes.

HermesIndices

Auto-generated bindings for index of the exports of hermes.

HermesPre

Auto-generated bindings for a pre-instantiated version of a component which implements the world hermes.

LinkOptions

Link-time configurations.