hermes::runtime_extensions::bindings::hermes::cardano::api

Trait HostBlock

pub trait HostBlock {
    // Required methods
    fn is_immutable(&mut self, self_: Resource<Block>) -> Result<bool>;
    fn is_rollback(
        &mut self,
        self_: Resource<Block>,
    ) -> Result<Result<bool, BlockError>>;
    fn get_txn(
        &mut self,
        self_: Resource<Block>,
        index: u16,
    ) -> Result<Result<Resource<Transaction>, TransactionError>>;
    fn get_slot(&mut self, self_: Resource<Block>) -> Result<u64>;
    fn raw(&mut self, self_: Resource<Block>) -> Result<Vec<u8>>;
    fn get_fork(&mut self, self_: Resource<Block>) -> Result<u64>;
    fn drop(&mut self, rep: Resource<Block>) -> Result<()>;
}

Required Methods

fn is_immutable(&mut self, self_: Resource<Block>) -> Result<bool>

Returns whether the block is part of the immutable section of the chain.

Returns

true if the block is in the immutable part.

false if the block is in the mutable part.

fn is_rollback( &mut self, self_: Resource<Block>, ) -> Result<Result<bool, BlockError>>

Returns whether the block is the first block of a rollback.

Returns

ok(bool) True if the block is the first block of a rollback, otherwise, False.

error(block-error): If block cannot be retrieved.

fn get_txn( &mut self, self_: Resource<Block>, index: u16, ) -> Result<Result<Resource<Transaction>, TransactionError>>

Retrieves a transaction at the specified index within the block.

Parameters

index : The index of the transaction to retrieve.

Returns

ok(transaction) : A transaction resource at the given index

error(transaction-error): If a transaction data does not exist in the block at the given index.

fn get_slot(&mut self, self_: Resource<Block>) -> Result<u64>

Retrieves the slot number that this block belongs to.

Returns

slot : The slot number of the block.

fn raw(&mut self, self_: Resource<Block>) -> Result<Vec<u8>>

Returns the raw CBOR representation of the block.

Returns

cbor : The CBOR format of the block.

fn get_fork(&mut self, self_: Resource<Block>) -> Result<u64>

Fork count is a counter that is incremented every time there is a roll-back in live-chain. It is used to help followers determine how far to roll-back to resynchronize without storing full block history. The fork count starts at 1 for live blocks and increments if the live chain tip is purged due to a detected fork, but it does not track the exact number of forks reported by peers.

0 - for all immutable data

1 - for any data read from the blockchain during a backfill on initial sync

2+ - for each subsequent rollback detected while reading live blocks.

Note: This fork terminology is different from fork in blockchain.

** Returns **

u64 : The fork count.

fn drop(&mut self, rep: Resource<Block>) -> Result<()>