hermes::runtime_extensions::bindings::hermes::cardano::api

Trait HostNetwork

pub trait HostNetwork {
    // Required methods
    fn new(
        &mut self,
        network: CardanoNetwork,
    ) -> Result<Result<Resource<Network>, CreateNetworkError>>;
    fn subscribe_block(
        &mut self,
        self_: Resource<Network>,
        start: SyncSlot,
    ) -> Result<Result<u32, SubscribeError>>;
    fn subscribe_immutable_roll_forward(
        &mut self,
        self_: Resource<Network>,
        start: SyncSlot,
    ) -> Result<Result<u32, SubscribeError>>;
    fn get_block(
        &mut self,
        self_: Resource<Network>,
        start: Option<u64>,
        step: i64,
    ) -> Result<Option<Resource<Block>>>;
    fn get_tips(
        &mut self,
        self_: Resource<Network>,
    ) -> Result<Option<(u64, u64)>>;
    fn drop(&mut self, rep: Resource<Network>) -> Result<()>;
}

Required Methods

fn new( &mut self, network: CardanoNetwork, ) -> Result<Result<Resource<Network>, CreateNetworkError>>

Create a new Cardano network instance.

Parameters

network: The Cardano network to connect to (e.g., Mainnet, Preprod, Preview).

Returns

ok(network): A resource network, if successfully create network resource.

error(create-network-error): If creating network resource failed.

fn subscribe_block( &mut self, self_: Resource<Network>, start: SyncSlot, ) -> Result<Result<u32, SubscribeError>>

Subscribe to blockchain block events, start from a specified starting point.

This sets up a subscription to receive new block and block rollback updates starting from the given start.

Parameters

start: The slot to begin following from.

Returns

ok(u32): A unsigned integer represent the underlying 32-bit representation of subscription ID resource. this subscription. Use to distinguishes events from different subscribers and provides control over subscription management.The ID must be unique across all active subscriptions.

error(subscribe-error): If subscription failed.

fn subscribe_immutable_roll_forward( &mut self, self_: Resource<Network>, start: SyncSlot, ) -> Result<Result<u32, SubscribeError>>

Subscribe to blockchain immutable rolls forward.

This sets up a subscription to receive event when the immutable part of the blockchain roll forwards.

Parameters

start: The slot to begin following from.

Returns

ok(u32): A unsigned integer represent the underlying 32-bit representation of subscription ID resource. this subscription. Use to distinguishes events from different subscribers and provides control over subscription management.The ID must be unique across all active subscriptions.

error(subscribe-error): If subscription failed.

fn get_block( &mut self, self_: Resource<Network>, start: Option<u64>, step: i64, ) -> Result<Option<Resource<Block>>>

Get a block relative to start by step.

Parameters

start: Slot to begin retrieval from, current tip if None.

step -0 : the block at start, will return None if there is no block exactly at this start slot. -+n: the n‑th block after the given start slot. –‑n: the n‑th block before the given start slot.

Note: For both +n and -n, the start does not need to be a true block. They will return the block which appears at this block offset, given the arbitrary start point. IF the start block does exist, it will never returned with a positive or negative step, as it is step 0.

Example, Given three consecutive blocks at slots 100, 200 and 300 the following will be returned: - start = 100, step = 0 -> 100 (Exact match) - start = 100, step = 2 -> 300 (Skips 200) - start = 150, step = 1 -> 200 (Rounds up from 150) - start = 200, step = 1 -> 300 (Forward iteration) - start = 300, step = -2 -> 100 (Skips 200) - start = 250, step = -2 -> 100 (Rounds down to 200 first)

Returns

Returns a block resource, None if block cannot be retrieved.

fn get_tips(&mut self, self_: Resource<Network>) -> Result<Option<(u64, u64)>>

Retrieve the current tips of the blockchain.

Returns

A tuple of two slots:

The immutable tip.

The mutable tip. None if the tips cannot be retrieved.

fn drop(&mut self, rep: Resource<Network>) -> Result<()>