hermes::runtime_extensions::bindings::hermes::cron
Module api

Structs

CronTagged

A tagged crontab entry It is valid for multiple crontab entries at the same time to have different tags. It is valid for crontab entries at different times to have the same tag. BUT there can only ever be 1 crontab entry at a specified time with a specified tag. ie, when + tag is uniquely identifying of every crontab entry. See: crontab.5 man page for details on cron schedule format.

Enums

CronComponent

A discreet time entry used to help convert numeric times into crontab entries.

Traits

Host

Functions

add_to_linker

Type Aliases

CronEventTag

A Tag used to mark a delivered cron event.

CronSched

A cron schedule in crontab format.

CronTime

A list of cron time components

Instant