Trait hermes::runtime_extensions::bindings::hermes::crypto::api::Host

pub trait Host: HostBip32Ed25519 {
    // Required method
    fn generate_mnemonic(
        &mut self,
        size: u8,
        prefix: Vec<String>,
        language: Option<String>
    ) -> Result<Result<Vec<String>, Errno>>;
}

Required Methods

fn generate_mnemonic( &mut self, size: u8, prefix: Vec<String>, language: Option<String> ) -> Result<Result<Vec<String>, Errno>>

Generate BIP39 Mnemonic Function

Generate a new BIP39 mnemonic phrase with the given size, prefix and language.

Parameters

size : The size of the mnemonic. Must be a multiple of 3 and in the range of 12 - 24. prefix : The prefix for the mnemonic. Must be a list of 1 - 3 words. language : Optional. The language to use for the mnemonic. If not provided, the default language is used.

Returns

Either a list of mnemonic words.

Or an error if the mnemonic could not be generated:

prefix-too-long : The prefix is longer than the maximum allowed length, max is 3.

invalid-mnemonic-length : The mnemonic length is not a multiple of 3 or not in the range of 12 - 24.

word-not-found : A word in the mnemonic is not found in the word list.

generate-entropy-failed : Failed to generate entropy.