hermes::runtime_extensions::bindings::hermes::crypto::api
Trait HostBip32Ed25519

pub trait HostBip32Ed25519 {
    // Required methods
    fn new(
        &mut self,
        mnemonic: Vec<String>,
        passphrase: Option<Vec<String>>,
    ) -> Result<Resource<Bip32Ed25519>>;
    fn public_key(
        &mut self,
        self_: Resource<Bip32Ed25519>,
    ) -> Result<(u64, u64, u64, u64)>;
    fn sign_data(
        &mut self,
        self_: Resource<Bip32Ed25519>,
        data: Vec<u8>,
    ) -> Result<(u64, u64, u64, u64, u64, u64, u64, u64)>;
    fn check_sig(
        &mut self,
        self_: Resource<Bip32Ed25519>,
        data: Vec<u8>,
        sig: (u64, u64, u64, u64, u64, u64, u64, u64),
    ) -> Result<bool>;
    fn derive(
        &mut self,
        self_: Resource<Bip32Ed25519>,
        path: String,
    ) -> Result<Resource<Bip32Ed25519>>;
    fn drop(&mut self, rep: Resource<Bip32Ed25519>) -> Result<()>;
}

Required Methods

fn new( &mut self, mnemonic: Vec<String>, passphrase: Option<Vec<String>>, ) -> Result<Resource<Bip32Ed25519>>

Create a new BIP32-Ed25519 Crypto resource

Parameters

mnemonic-phrase : BIP39 mnemonic.

passphrase : Optional BIP39 passphrase.

fn public_key( &mut self, self_: Resource<Bip32Ed25519>, ) -> Result<(u64, u64, u64, u64)>

Get the public key for this private key.

fn sign_data( &mut self, self_: Resource<Bip32Ed25519>, data: Vec<u8>, ) -> Result<(u64, u64, u64, u64, u64, u64, u64, u64)>

Sign data with the Private key, and return it.

Parameters

data : The data to sign.

fn check_sig( &mut self, self_: Resource<Bip32Ed25519>, data: Vec<u8>, sig: (u64, u64, u64, u64, u64, u64, u64, u64), ) -> Result<bool>

Check a signature on a set of data.

Parameters

data : The data to check.

sig : The signature to check.

Returns

true : Signature checked OK.

false : Signature check failed.

fn derive( &mut self, self_: Resource<Bip32Ed25519>, path: String, ) -> Result<Resource<Bip32Ed25519>>

Derive a new private key from the current private key.

Parameters

path : Derivation path.

Note: uses BIP32 HD key derivation.

fn drop(&mut self, rep: Resource<Bip32Ed25519>) -> Result<()>