Trait hermes::runtime_extensions::bindings::hermes::crypto::api::HostEd25519Bip32

pub trait HostEd25519Bip32 {
    // Required methods
    fn new(
        &mut self,
        private_key: Option<(u64, u64, u64, u64)>
    ) -> Result<Resource<Ed25519Bip32>>;
    fn public_key(
        &mut self,
        self_: Resource<Ed25519Bip32>
    ) -> Result<(u64, u64, u64, u64)>;
    fn sign_data(
        &mut self,
        self_: Resource<Ed25519Bip32>,
        data: Vec<u8>
    ) -> Result<(u64, u64, u64, u64)>;
    fn check_sig(
        &mut self,
        self_: Resource<Ed25519Bip32>,
        data: Vec<u8>,
        sig: (u64, u64, u64, u64)
    ) -> Result<bool>;
    fn derive(
        &mut self,
        self_: Resource<Ed25519Bip32>
    ) -> Result<Resource<Ed25519Bip32>>;
    fn gen_private_key(&mut self) -> Result<(u64, u64, u64, u64)>;
    fn drop(&mut self, rep: Resource<Ed25519Bip32>) -> Result<()>;
}

Required Methods

fn new( &mut self, private_key: Option<(u64, u64, u64, u64)> ) -> Result<Resource<Ed25519Bip32>>

Create a new ED25519-BIP32 Crypto resource

Parameters

private_key : The key to use, if not supplied one is RANDOMLY generated.

fn public_key( &mut self, self_: Resource<Ed25519Bip32> ) -> Result<(u64, u64, u64, u64)>

Get the public key for this private key.

fn sign_data( &mut self, self_: Resource<Ed25519Bip32>, data: Vec<u8> ) -> Result<(u64, u64, u64, u64)>

Sign data with the Private key, and return it.

Parameters

data : The data to sign.

fn check_sig( &mut self, self_: Resource<Ed25519Bip32>, data: Vec<u8>, sig: (u64, u64, u64, u64) ) -> Result<bool>

Check a signature on a set of data.

Parameters

data : The data to check.

sig : The signature to check.

Returns

true : Signature checked OK.

false : Signature check failed.

fn derive( &mut self, self_: Resource<Ed25519Bip32> ) -> Result<Resource<Ed25519Bip32>>

Derive a new private key from the current private key.

Note: uses BIP32 HD key derivation.

fn gen_private_key(&mut self) -> Result<(u64, u64, u64, u64)>

Create a new RANDOM private key.

Note, this does not need to be used, as the constructor will do this automatically.

fn drop(&mut self, rep: Resource<Ed25519Bip32>) -> Result<()>