Trait hermes::runtime_extensions::bindings::hermes::sqlite::api::Host

pub trait Host: HostSqlite + HostStatement {
    // Required method
    fn open(
        &mut self,
        readonly: bool,
        memory: bool
    ) -> Result<Result<Resource<Sqlite>, Errno>>;
}

Required Methods

fn open( &mut self, readonly: bool, memory: bool ) -> Result<Result<Resource<Sqlite>, Errno>>

Opens a connection to a new or existing SQLite database.

Parameters

readonly: If set to true, the database is opened in read-only mode. An error is returned if the database doesn’t already exist.

memory: If set to true, the database will be opened as an in-memory database.

Returns

If the database is opened (and/or created) successfully, then the sqlite3 object is returned. Otherwise an error code is returned.