hermes::runtime_extensions::bindings::hermes::sqlite::api

Trait HostSqlite

pub trait HostSqlite: Sized {
    // Required methods
    fn close(&mut self, self_: Resource<Sqlite>) -> Result<Result<(), Errno>>;
    fn errcode(&mut self, self_: Resource<Sqlite>) -> Result<Option<ErrorInfo>>;
    fn prepare(
        &mut self,
        self_: Resource<Sqlite>,
        sql: String,
    ) -> Result<Result<Resource<Statement>, Errno>>;
    fn execute(
        &mut self,
        self_: Resource<Sqlite>,
        sql: String,
    ) -> Result<Result<(), Errno>>;
    fn drop(&mut self, rep: Resource<Sqlite>) -> Result<()>;
}

Required Methods

fn close(&mut self, self_: Resource<Sqlite>) -> Result<Result<(), Errno>>

Closes a database connection, destructor for sqlite3.

Ideally, applications should finalize all prepared statements associated with the sqlite3 object prior to attempting to close the object. If the database connection is associated with unfinalized prepared statements, then the function will leave the database connection open and return the busy error code.

If an sqlite3 object is destroyed while a transaction is open, the transaction is automatically rolled back.

fn errcode(&mut self, self_: Resource<Sqlite>) -> Result<Option<ErrorInfo>>

Retrieves the numeric result code for the most recent failed SQLite operation on a database connection.

Returns

The error object containing numeric code and detail message for the most recent failed SQLite operation. If there is no recent failed, none is returned.

fn prepare( &mut self, self_: Resource<Sqlite>, sql: String, ) -> Result<Result<Resource<Statement>, Errno>>

Compiles SQL text into byte-code that will do the work of querying or updating the database.

Parameters

db: Database handle.

sql: SQL statement, UTF-8 encoded.

Returns

A compiled prepared statement that can be executed using sqlite3_step(). If there is an error or the input text contains no SQL (if the input is an empty string or a comment) then an error code is returned.

fn execute( &mut self, self_: Resource<Sqlite>, sql: String, ) -> Result<Result<(), Errno>>

Executes an SQL query directly without preparing it into a statement and returns the result.

Parameters

sql: SQL statement, UTF-8 encoded.

fn drop(&mut self, rep: Resource<Sqlite>) -> Result<()>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.