hermes::runtime_extensions::bindings::hermes::sqlite::api

Trait HostStatement

pub trait HostStatement {
    // Required methods
    fn bind(
        &mut self,
        self_: Resource<Statement>,
        index: u32,
        value: Value,
    ) -> Result<Result<(), Errno>>;
    fn step(&mut self, self_: Resource<Statement>) -> Result<Result<(), Errno>>;
    fn column(
        &mut self,
        self_: Resource<Statement>,
        index: u32,
    ) -> Result<Result<Value, Errno>>;
    fn finalize(
        &mut self,
        self_: Resource<Statement>,
    ) -> Result<Result<(), Errno>>;
    fn drop(&mut self, rep: Resource<Statement>) -> Result<()>;
}

Required Methods

fn bind( &mut self, self_: Resource<Statement>, index: u32, value: Value, ) -> Result<Result<(), Errno>>

Stores application data into parameters of the original SQL.

Parameters

index: The index of the SQL parameter to be set.

value: The value to bind to the parameter.

fn step(&mut self, self_: Resource<Statement>) -> Result<Result<(), Errno>>

Advances a statement to the next result row or to completion.

After a prepared statement has been prepared, this function must be called one or more times to evaluate the statement.

fn column( &mut self, self_: Resource<Statement>, index: u32, ) -> Result<Result<Value, Errno>>

Returns information about a single column of the current result row of a query.

If the SQL statement does not currently point to a valid row, or if the column index is out of range, the result is undefined.

Parameters

index: The index of the column for which information should be returned. The leftmost column of the result set has the index 0.

Returns

The value of a result column in a specific data format.

fn finalize(&mut self, self_: Resource<Statement>) -> Result<Result<(), Errno>>

Destroys a prepared statement object. If the most recent evaluation of the statement encountered no errors or if the statement is never been evaluated, then the function results without errors. If the most recent evaluation of statement failed, then the function results the appropriate error code.

The application must finalize every prepared statement in order to avoid resource leaks. It is a grievous error for the application to try to use a prepared statement after it has been finalized. Any use of a prepared statement after it has been finalized can result in undefined and undesirable behavior such as segfaults and heap corruption.

fn drop(&mut self, rep: Resource<Statement>) -> Result<()>