hermes::runtime_extensions::bindings::wasi::clocks::monotonic_clock

Trait Host

pub trait Host {
    // Required methods
    fn now(&mut self) -> Result<u64>;
    fn resolution(&mut self) -> Result<u64>;
    fn subscribe_instant(&mut self, when: u64) -> Result<Resource<Pollable>>;
    fn subscribe_duration(&mut self, when: u64) -> Result<Resource<Pollable>>;
}

Required Methods

fn now(&mut self) -> Result<u64>

Read the current value of the clock.

The clock is monotonic, therefore calling this function repeatedly will produce a sequence of non-decreasing values.

fn resolution(&mut self) -> Result<u64>

Query the resolution of the clock. Returns the duration of time corresponding to a clock tick.

fn subscribe_instant(&mut self, when: u64) -> Result<Resource<Pollable>>

Create a pollable which will resolve once the specified instant has occurred.

fn subscribe_duration(&mut self, when: u64) -> Result<Resource<Pollable>>

Create a pollable that will resolve after the specified duration has elapsed from the time this function is invoked.