hermes::runtime_extensions::bindings::wasi::filesystem
Module types

Structs

DescriptorFlags

DescriptorStat

File attributes.

DirectoryEntry

A directory entry.

MetadataHashValue

A 128-bit hash value, split into parts because wasm doesn’t have a 128-bit integer type.

OpenFlags

PathFlags

Enums

Advice

File or memory access pattern advisory information.

Descriptor

A descriptor is a reference to a filesystem object, which may be a file, directory, named pipe, special file, or other object on which filesystem calls may be made.

DescriptorType

The type of a filesystem object referenced by a descriptor.

DirectoryEntryStream

A stream of directory entries.

ErrorCode

Error codes returned by functions, similar to errno in POSIX. Not all of these error codes are returned by the functions provided by this API; some are used in higher-level library layers, and others are provided merely for alignment with POSIX.

NewTimestamp

When setting a timestamp, this gives the value to set it to.

Traits

Host

HostDescriptor

HostDirectoryEntryStream

Functions

add_to_linker

Type Aliases

Datetime

Error

Filesize

File size or length of a region within a file.

InputStream

LinkCount

Number of hard links to an inode.

OutputStream