Enum hermes::runtime_extensions::bindings::wasi::filesystem::types::ErrorCode

pub enum ErrorCode {

    Access,
    WouldBlock,
    Already,
    BadDescriptor,
    Busy,
    Deadlock,
    Quota,
    Exist,
    FileTooLarge,
    IllegalByteSequence,
    InProgress,
    Interrupted,
    Invalid,
    Io,
    IsDirectory,
    Loop,
    TooManyLinks,
    MessageSize,
    NameTooLong,
    NoDevice,
    NoEntry,
    NoLock,
    InsufficientMemory,
    InsufficientSpace,
    NotDirectory,
    NotEmpty,
    NotRecoverable,
    Unsupported,
    NoTty,
    NoSuchDevice,
    Overflow,
    NotPermitted,
    Pipe,
    ReadOnly,
    InvalidSeek,
    TextFileBusy,
    CrossDevice,

}

Error codes returned by functions, similar to errno in POSIX. Not all of these error codes are returned by the functions provided by this API; some are used in higher-level library layers, and others are provided merely for alignment with POSIX.

Variants

Access

Permission denied, similar to EACCES in POSIX.

WouldBlock

Resource unavailable, or operation would block, similar to EAGAIN and EWOULDBLOCK in POSIX.

Already

Connection already in progress, similar to EALREADY in POSIX.

BadDescriptor

Bad descriptor, similar to EBADF in POSIX.

Busy

Device or resource busy, similar to EBUSY in POSIX.

Deadlock

Resource deadlock would occur, similar to EDEADLK in POSIX.

Quota

Storage quota exceeded, similar to EDQUOT in POSIX.

Exist

File exists, similar to EEXIST in POSIX.

FileTooLarge

File too large, similar to EFBIG in POSIX.

IllegalByteSequence

Illegal byte sequence, similar to EILSEQ in POSIX.

InProgress

Operation in progress, similar to EINPROGRESS in POSIX.

Interrupted

Interrupted function, similar to EINTR in POSIX.

Invalid

Invalid argument, similar to EINVAL in POSIX.

Io

I/O error, similar to EIO in POSIX.

IsDirectory

Is a directory, similar to EISDIR in POSIX.

Loop

Too many levels of symbolic links, similar to ELOOP in POSIX.

TooManyLinks

Too many links, similar to EMLINK in POSIX.

MessageSize

Message too large, similar to EMSGSIZE in POSIX.

NameTooLong

Filename too long, similar to ENAMETOOLONG in POSIX.

NoDevice

No such device, similar to ENODEV in POSIX.

NoEntry

No such file or directory, similar to ENOENT in POSIX.

NoLock

No locks available, similar to ENOLCK in POSIX.

InsufficientMemory

Not enough space, similar to ENOMEM in POSIX.

InsufficientSpace

No space left on device, similar to ENOSPC in POSIX.

NotDirectory

Not a directory or a symbolic link to a directory, similar to ENOTDIR in POSIX.

NotEmpty

Directory not empty, similar to ENOTEMPTY in POSIX.

NotRecoverable

State not recoverable, similar to ENOTRECOVERABLE in POSIX.

Unsupported

Not supported, similar to ENOTSUP and ENOSYS in POSIX.

NoTty

Inappropriate I/O control operation, similar to ENOTTY in POSIX.

NoSuchDevice

No such device or address, similar to ENXIO in POSIX.

Overflow

Value too large to be stored in data type, similar to EOVERFLOW in POSIX.

NotPermitted

Operation not permitted, similar to EPERM in POSIX.

Pipe

Broken pipe, similar to EPIPE in POSIX.

ReadOnly

Read-only file system, similar to EROFS in POSIX.

InvalidSeek

Invalid seek, similar to ESPIPE in POSIX.

TextFileBusy

Text file busy, similar to ETXTBSY in POSIX.

CrossDevice

Cross-device link, similar to EXDEV in POSIX.

Implementations

impl ErrorCode

pub fn name(&self) -> &'static str

pub fn message(&self) -> &'static str