hermes::runtime_extensions::bindings::wasi::filesystem::types

Trait HostDescriptor

pub trait HostDescriptor {

    // Required methods
    fn read_via_stream(
        &mut self,
        self_: Resource<Descriptor>,
        offset: u64,
    ) -> Result<Result<Resource<InputStream>, ErrorCode>>;
    fn write_via_stream(
        &mut self,
        self_: Resource<Descriptor>,
        offset: u64,
    ) -> Result<Result<Resource<OutputStream>, ErrorCode>>;
    fn append_via_stream(
        &mut self,
        self_: Resource<Descriptor>,
    ) -> Result<Result<Resource<OutputStream>, ErrorCode>>;
    fn advise(
        &mut self,
        self_: Resource<Descriptor>,
        offset: u64,
        length: u64,
        advice: Advice,
    ) -> Result<Result<(), ErrorCode>>;
    fn sync_data(
        &mut self,
        self_: Resource<Descriptor>,
    ) -> Result<Result<(), ErrorCode>>;
    fn get_flags(
        &mut self,
        self_: Resource<Descriptor>,
    ) -> Result<Result<DescriptorFlags, ErrorCode>>;
    fn get_type(
        &mut self,
        self_: Resource<Descriptor>,
    ) -> Result<Result<DescriptorType, ErrorCode>>;
    fn set_size(
        &mut self,
        self_: Resource<Descriptor>,
        size: u64,
    ) -> Result<Result<(), ErrorCode>>;
    fn set_times(
        &mut self,
        self_: Resource<Descriptor>,
        data_access_timestamp: NewTimestamp,
        data_modification_timestamp: NewTimestamp,
    ) -> Result<Result<(), ErrorCode>>;
    fn read(
        &mut self,
        self_: Resource<Descriptor>,
        length: u64,
        offset: u64,
    ) -> Result<Result<(Vec<u8>, bool), ErrorCode>>;
    fn write(
        &mut self,
        self_: Resource<Descriptor>,
        buffer: Vec<u8>,
        offset: u64,
    ) -> Result<Result<u64, ErrorCode>>;
    fn read_directory(
        &mut self,
        self_: Resource<Descriptor>,
    ) -> Result<Result<Resource<DirectoryEntryStream>, ErrorCode>>;
    fn sync(
        &mut self,
        self_: Resource<Descriptor>,
    ) -> Result<Result<(), ErrorCode>>;
    fn create_directory_at(
        &mut self,
        self_: Resource<Descriptor>,
        path: String,
    ) -> Result<Result<(), ErrorCode>>;
    fn stat(
        &mut self,
        self_: Resource<Descriptor>,
    ) -> Result<Result<DescriptorStat, ErrorCode>>;
    fn stat_at(
        &mut self,
        self_: Resource<Descriptor>,
        path_flags: PathFlags,
        path: String,
    ) -> Result<Result<DescriptorStat, ErrorCode>>;
    fn set_times_at(
        &mut self,
        self_: Resource<Descriptor>,
        path_flags: PathFlags,
        path: String,
        data_access_timestamp: NewTimestamp,
        data_modification_timestamp: NewTimestamp,
    ) -> Result<Result<(), ErrorCode>>;
    fn link_at(
        &mut self,
        self_: Resource<Descriptor>,
        old_path_flags: PathFlags,
        old_path: String,
        new_descriptor: Resource<Descriptor>,
        new_path: String,
    ) -> Result<Result<(), ErrorCode>>;
    fn open_at(
        &mut self,
        self_: Resource<Descriptor>,
        path_flags: PathFlags,
        path: String,
        open_flags: OpenFlags,
        flags: DescriptorFlags,
    ) -> Result<Result<Resource<Descriptor>, ErrorCode>>;
    fn readlink_at(
        &mut self,
        self_: Resource<Descriptor>,
        path: String,
    ) -> Result<Result<String, ErrorCode>>;
    fn remove_directory_at(
        &mut self,
        self_: Resource<Descriptor>,
        path: String,
    ) -> Result<Result<(), ErrorCode>>;
    fn rename_at(
        &mut self,
        self_: Resource<Descriptor>,
        old_path: String,
        new_descriptor: Resource<Descriptor>,
        new_path: String,
    ) -> Result<Result<(), ErrorCode>>;
    fn symlink_at(
        &mut self,
        self_: Resource<Descriptor>,
        old_path: String,
        new_path: String,
    ) -> Result<Result<(), ErrorCode>>;
    fn unlink_file_at(
        &mut self,
        self_: Resource<Descriptor>,
        path: String,
    ) -> Result<Result<(), ErrorCode>>;
    fn is_same_object(
        &mut self,
        self_: Resource<Descriptor>,
        other: Resource<Descriptor>,
    ) -> Result<bool>;
    fn metadata_hash(
        &mut self,
        self_: Resource<Descriptor>,
    ) -> Result<Result<MetadataHashValue, ErrorCode>>;
    fn metadata_hash_at(
        &mut self,
        self_: Resource<Descriptor>,
        path_flags: PathFlags,
        path: String,
    ) -> Result<Result<MetadataHashValue, ErrorCode>>;
    fn drop(&mut self, rep: Resource<Descriptor>) -> Result<()>;

}

Required Methods

fn read_via_stream( &mut self, self_: Resource<Descriptor>, offset: u64, ) -> Result<Result<Resource<InputStream>, ErrorCode>>

Return a stream for reading from a file, if available.

May fail with an error-code describing why the file cannot be read.

Multiple read, write, and append streams may be active on the same open file and they do not interfere with each other.

Note: This allows using read-stream, which is similar to read in POSIX.

fn write_via_stream( &mut self, self_: Resource<Descriptor>, offset: u64, ) -> Result<Result<Resource<OutputStream>, ErrorCode>>

Return a stream for writing to a file, if available.

May fail with an error-code describing why the file cannot be written.

Note: This allows using write-stream, which is similar to write in POSIX.

fn append_via_stream( &mut self, self_: Resource<Descriptor>, ) -> Result<Result<Resource<OutputStream>, ErrorCode>>

Return a stream for appending to a file, if available.

May fail with an error-code describing why the file cannot be appended.

Note: This allows using write-stream, which is similar to write with O_APPEND in POSIX.

fn advise( &mut self, self_: Resource<Descriptor>, offset: u64, length: u64, advice: Advice, ) -> Result<Result<(), ErrorCode>>

Provide file advisory information on a descriptor.

This is similar to posix_fadvise in POSIX.

fn sync_data( &mut self, self_: Resource<Descriptor>, ) -> Result<Result<(), ErrorCode>>

Synchronize the data of a file to disk.

This function succeeds with no effect if the file descriptor is not opened for writing.

Note: This is similar to fdatasync in POSIX.

fn get_flags( &mut self, self_: Resource<Descriptor>, ) -> Result<Result<DescriptorFlags, ErrorCode>>

Get flags associated with a descriptor.

Note: This returns similar flags to fcntl(fd, F_GETFL) in POSIX.

Note: This returns the value that was the fs_flags value returned from fdstat_get in earlier versions of WASI.

fn get_type( &mut self, self_: Resource<Descriptor>, ) -> Result<Result<DescriptorType, ErrorCode>>

Get the dynamic type of a descriptor.

Note: This returns the same value as the type field of the fd-stat returned by stat, stat-at and similar.

Note: This returns similar flags to the st_mode & S_IFMT value provided by fstat in POSIX.

Note: This returns the value that was the fs_filetype value returned from fdstat_get in earlier versions of WASI.

fn set_size( &mut self, self_: Resource<Descriptor>, size: u64, ) -> Result<Result<(), ErrorCode>>

Adjust the size of an open file. If this increases the file’s size, the extra bytes are filled with zeros.

Note: This was called fd_filestat_set_size in earlier versions of WASI.

fn set_times( &mut self, self_: Resource<Descriptor>, data_access_timestamp: NewTimestamp, data_modification_timestamp: NewTimestamp, ) -> Result<Result<(), ErrorCode>>

Adjust the timestamps of an open file or directory.

Note: This is similar to futimens in POSIX.

Note: This was called fd_filestat_set_times in earlier versions of WASI.

fn read( &mut self, self_: Resource<Descriptor>, length: u64, offset: u64, ) -> Result<Result<(Vec<u8>, bool), ErrorCode>>

Read from a descriptor, without using and updating the descriptor’s offset.

This function returns a list of bytes containing the data that was read, along with a bool which, when true, indicates that the end of the file was reached. The returned list will contain up to length bytes; it may return fewer than requested, if the end of the file is reached or if the I/O operation is interrupted.

In the future, this may change to return a stream<u8, error-code>.

Note: This is similar to pread in POSIX.

fn write( &mut self, self_: Resource<Descriptor>, buffer: Vec<u8>, offset: u64, ) -> Result<Result<u64, ErrorCode>>

Write to a descriptor, without using and updating the descriptor’s offset.

It is valid to write past the end of a file; the file is extended to the extent of the write, with bytes between the previous end and the start of the write set to zero.

In the future, this may change to take a stream<u8, error-code>.

Note: This is similar to pwrite in POSIX.

fn read_directory( &mut self, self_: Resource<Descriptor>, ) -> Result<Result<Resource<DirectoryEntryStream>, ErrorCode>>

Read directory entries from a directory.

On filesystems where directories contain entries referring to themselves and their parents, often named . and .. respectively, these entries are omitted.

This always returns a new stream which starts at the beginning of the directory. Multiple streams may be active on the same directory, and they do not interfere with each other.

fn sync(&mut self, self_: Resource<Descriptor>) -> Result<Result<(), ErrorCode>>

Synchronize the data and metadata of a file to disk.

This function succeeds with no effect if the file descriptor is not opened for writing.

Note: This is similar to fsync in POSIX.

fn create_directory_at( &mut self, self_: Resource<Descriptor>, path: String, ) -> Result<Result<(), ErrorCode>>

Create a directory.

Note: This is similar to mkdirat in POSIX.

fn stat( &mut self, self_: Resource<Descriptor>, ) -> Result<Result<DescriptorStat, ErrorCode>>

Return the attributes of an open file or directory.

Note: This is similar to fstat in POSIX, except that it does not return device and inode information. For testing whether two descriptors refer to the same underlying filesystem object, use is-same-object. To obtain additional data that can be used do determine whether a file has been modified, use metadata-hash.

Note: This was called fd_filestat_get in earlier versions of WASI.

fn stat_at( &mut self, self_: Resource<Descriptor>, path_flags: PathFlags, path: String, ) -> Result<Result<DescriptorStat, ErrorCode>>

Return the attributes of a file or directory.

Note: This is similar to fstatat in POSIX, except that it does not return device and inode information. See the stat description for a discussion of alternatives.

Note: This was called path_filestat_get in earlier versions of WASI.

fn set_times_at( &mut self, self_: Resource<Descriptor>, path_flags: PathFlags, path: String, data_access_timestamp: NewTimestamp, data_modification_timestamp: NewTimestamp, ) -> Result<Result<(), ErrorCode>>

Adjust the timestamps of a file or directory.

Note: This is similar to utimensat in POSIX.

Note: This was called path_filestat_set_times in earlier versions of WASI.

fn link_at( &mut self, self_: Resource<Descriptor>, old_path_flags: PathFlags, old_path: String, new_descriptor: Resource<Descriptor>, new_path: String, ) -> Result<Result<(), ErrorCode>>

Create a hard link.

Fails with error-code::no-entry if the old path does not exist, with error-code::exist if the new path already exists, and error-code::not-permitted if the old path is not a file.

Note: This is similar to linkat in POSIX.

fn open_at( &mut self, self_: Resource<Descriptor>, path_flags: PathFlags, path: String, open_flags: OpenFlags, flags: DescriptorFlags, ) -> Result<Result<Resource<Descriptor>, ErrorCode>>

Open a file or directory.

If flags contains descriptor-flags::mutate-directory, and the base descriptor doesn’t have descriptor-flags::mutate-directory set, open-at fails with error-code::read-only.

If flags contains write or mutate-directory, or open-flags contains truncate or create, and the base descriptor doesn’t have descriptor-flags::mutate-directory set, open-at fails with error-code::read-only.

Note: This is similar to openat in POSIX.

fn readlink_at( &mut self, self_: Resource<Descriptor>, path: String, ) -> Result<Result<String, ErrorCode>>

Read the contents of a symbolic link.

If the contents contain an absolute or rooted path in the underlying filesystem, this function fails with error-code::not-permitted.

Note: This is similar to readlinkat in POSIX.

fn remove_directory_at( &mut self, self_: Resource<Descriptor>, path: String, ) -> Result<Result<(), ErrorCode>>

Remove a directory.

Return error-code::not-empty if the directory is not empty.

Note: This is similar to unlinkat(fd, path, AT_REMOVEDIR) in POSIX.

fn rename_at( &mut self, self_: Resource<Descriptor>, old_path: String, new_descriptor: Resource<Descriptor>, new_path: String, ) -> Result<Result<(), ErrorCode>>

Rename a filesystem object.

Note: This is similar to renameat in POSIX.

fn symlink_at( &mut self, self_: Resource<Descriptor>, old_path: String, new_path: String, ) -> Result<Result<(), ErrorCode>>

Create a symbolic link (also known as a “symlink”).

If old-path starts with /, the function fails with error-code::not-permitted.

Note: This is similar to symlinkat in POSIX.

fn unlink_file_at( &mut self, self_: Resource<Descriptor>, path: String, ) -> Result<Result<(), ErrorCode>>

Unlink a filesystem object that is not a directory.

Return error-code::is-directory if the path refers to a directory. Note: This is similar to unlinkat(fd, path, 0) in POSIX.

fn is_same_object( &mut self, self_: Resource<Descriptor>, other: Resource<Descriptor>, ) -> Result<bool>

Test whether two descriptors refer to the same filesystem object.

In POSIX, this corresponds to testing whether the two descriptors have the same device (st_dev) and inode (st_ino or d_ino) numbers. wasi-filesystem does not expose device and inode numbers, so this function may be used instead.

fn metadata_hash( &mut self, self_: Resource<Descriptor>, ) -> Result<Result<MetadataHashValue, ErrorCode>>

Return a hash of the metadata associated with a filesystem object referred to by a descriptor.

This returns a hash of the last-modification timestamp and file size, and may also include the inode number, device number, birth timestamp, and other metadata fields that may change when the file is modified or replaced. It may also include a secret value chosen by the implementation and not otherwise exposed.

Implementations are encouraged to provide the following properties:

If the file is not modified or replaced, the computed hash value should usually not change.

If the object is modified or replaced, the computed hash value should usually change.

The inputs to the hash should not be easily computable from the computed hash.

However, none of these is required.

fn metadata_hash_at( &mut self, self_: Resource<Descriptor>, path_flags: PathFlags, path: String, ) -> Result<Result<MetadataHashValue, ErrorCode>>

Return a hash of the metadata associated with a filesystem object referred to by a directory descriptor and a relative path.

This performs the same hash computation as metadata-hash.

fn drop(&mut self, rep: Resource<Descriptor>) -> Result<()>