hermes::runtime_extensions::bindings::wasi::http::incoming_handler

Trait Host

pub trait Host {
    // Required method
    fn handle(
        &mut self,
        request: Resource<IncomingRequest>,
        response_out: Resource<ResponseOutparam>,
    ) -> Result<()>;
}

Required Methods

fn handle( &mut self, request: Resource<IncomingRequest>, response_out: Resource<ResponseOutparam>, ) -> Result<()>

This function is invoked with an incoming HTTP Request, and a resource response-outparam which provides the capability to reply with an HTTP Response. The response is sent by calling the response-outparam.set method, which allows execution to continue after the response has been sent. This enables both streaming to the response body, and performing other work.

The implementor of this function must write a response to the response-outparam before returning, or else the caller will respond with an error on its behalf.