hermes::runtime_extensions::bindings::wasi::http::types

Trait HostOutgoingResponse

pub trait HostOutgoingResponse {
    // Required methods
    fn new(
        &mut self,
        headers: Resource<Fields>,
    ) -> Result<Resource<OutgoingResponse>>;
    fn status_code(&mut self, self_: Resource<OutgoingResponse>) -> Result<u16>;
    fn set_status_code(
        &mut self,
        self_: Resource<OutgoingResponse>,
        status_code: u16,
    ) -> Result<Result<(), ()>>;
    fn headers(
        &mut self,
        self_: Resource<OutgoingResponse>,
    ) -> Result<Resource<Fields>>;
    fn body(
        &mut self,
        self_: Resource<OutgoingResponse>,
    ) -> Result<Result<Resource<OutgoingBody>, ()>>;
    fn drop(&mut self, rep: Resource<OutgoingResponse>) -> Result<()>;
}

Required Methods

fn new( &mut self, headers: Resource<Fields>, ) -> Result<Resource<OutgoingResponse>>

Construct an outgoing-response, with a default status-code of 200. If a different status-code is needed, it must be set via the set-status-code method.

headers is the HTTP Headers for the Response.

fn status_code(&mut self, self_: Resource<OutgoingResponse>) -> Result<u16>

Get the HTTP Status Code for the Response.

fn set_status_code( &mut self, self_: Resource<OutgoingResponse>, status_code: u16, ) -> Result<Result<(), ()>>

Set the HTTP Status Code for the Response. Fails if the status-code given is not a valid http status code.

fn headers( &mut self, self_: Resource<OutgoingResponse>, ) -> Result<Resource<Fields>>

Get the headers associated with the Request.

The returned headers resource is immutable: set, append, and delete operations will fail with header-error.immutable.

This headers resource is a child: it must be dropped before the parent outgoing-request is dropped, or its ownership is transfered to another component by e.g. outgoing-handler.handle.

fn body( &mut self, self_: Resource<OutgoingResponse>, ) -> Result<Result<Resource<OutgoingBody>, ()>>

Returns the resource corresponding to the outgoing Body for this Response.

Returns success on the first call: the outgoing-body resource for this outgoing-response can be retrieved at most once. Subsequent calls will return error.

fn drop(&mut self, rep: Resource<OutgoingResponse>) -> Result<()>