hermes::runtime_extensions::bindings::wasi::io::error

Trait HostError

pub trait HostError {
    // Required methods
    fn to_debug_string(&mut self, self_: Resource<Error>) -> Result<String>;
    fn drop(&mut self, rep: Resource<Error>) -> Result<()>;
}

Required Methods

fn to_debug_string(&mut self, self_: Resource<Error>) -> Result<String>

Returns a string that is suitable to assist humans in debugging this error.

WARNING: The returned string should not be consumed mechanically! It may change across platforms, hosts, or other implementation details. Parsing this string is a major platform-compatibility hazard.

fn drop(&mut self, rep: Resource<Error>) -> Result<()>