hermes::runtime_extensions::hermes::cardano::event

Trait ComponentInstanceExt

trait ComponentInstanceExt {
    // Required methods
    fn lookup_hermes_cardano_event_on_block_on_cardano_block(
        &self,
        store: &mut Store<HermesRuntimeContext>,
    ) -> Result<TypedFunc<(Resource<SubscriptionId>, Resource<Block>), ()>, Error>;
    fn lookup_hermes_cardano_event_on_immutable_roll_forward_on_cardano_immutable_roll_forward(
        &self,
        store: &mut Store<HermesRuntimeContext>,
    ) -> Result<TypedFunc<(Resource<SubscriptionId>, Resource<Block>), ()>, Error>;

    // Provided methods
    fn hermes_cardano_event_on_block_on_cardano_block(
        &self,
        store: &mut Store<HermesRuntimeContext>,
        subscription_id: Resource<SubscriptionId>,
        block_id: Resource<Block>,
    ) -> Result<()> { ... }
    fn hermes_cardano_event_on_immutable_roll_forward_on_cardano_immutable_roll_forward(
        &self,
        store: &mut Store<HermesRuntimeContext>,
        subscription_id: Resource<SubscriptionId>,
        block_id: Resource<Block>,
    ) -> Result<()> { ... }
}

Required Methods

fn lookup_hermes_cardano_event_on_block_on_cardano_block( &self, store: &mut Store<HermesRuntimeContext>, ) -> Result<TypedFunc<(Resource<SubscriptionId>, Resource<Block>), ()>, Error>

Looks up “on-cardano-block” from “hermes:cardano/event-on-block”.

Panic

Lookup never panics, calling the export can panic if misused (see [wasmtime::component::Func::call]).

fn lookup_hermes_cardano_event_on_immutable_roll_forward_on_cardano_immutable_roll_forward( &self, store: &mut Store<HermesRuntimeContext>, ) -> Result<TypedFunc<(Resource<SubscriptionId>, Resource<Block>), ()>, Error>

Looks up “on-immutable-roll-forward” from “hermes:cardano/event-on-immutable-roll-forward”.

Panic

Lookup never panics, calling the export can panic if misused (see [wasmtime::component::Func::call]).

Provided Methods

fn hermes_cardano_event_on_block_on_cardano_block( &self, store: &mut Store<HermesRuntimeContext>, subscription_id: Resource<SubscriptionId>, block_id: Resource<Block>, ) -> Result<()>

Looks up and calls “on-cardano-block” from “hermes:cardano/event-on-block”. This wraps around Self::lookup_hermes_cardano_event_on_block_on_cardano_block.

Panic

Calling the export can panic if misused (see [wasmtime::component::Func::call]).

fn hermes_cardano_event_on_immutable_roll_forward_on_cardano_immutable_roll_forward( &self, store: &mut Store<HermesRuntimeContext>, subscription_id: Resource<SubscriptionId>, block_id: Resource<Block>, ) -> Result<()>

Looks up and calls “on-immutable-roll-forward” from “hermes:cardano/event-on-immutable-roll-forward”. This wraps around Self::lookup_hermes_cardano_event_on_immutable_roll_forward_on_cardano_immutable_roll_forward.

Panic

Calling the export can panic if misused (see [wasmtime::component::Func::call]).