hermes::runtime_extensions::hermes::cardano::network

Function subscribe

async fn subscribe(
    cmd_rx: Receiver<Command>,
    app: ApplicationName,
    module_id: ModuleId,
    start: Point,
    network: Network,
    subscription_type: SubscriptionType,
    subscription_id: Resource<SubscriptionId>,
)

Subscribe to events from a Cardano network. This function will enter a loop and await either a command to stop the subscription or a new block update from the chain follower.