hermes::runtime_extensions::hermes::cardano::tokio_runtime_task
Function spawn_follower

async fn spawn_follower(
    app_name: ApplicationName,
    module_id: ModuleId,
    chain_id: CardanoBlockchainId,
    follow_from: PointOrTip,
) -> Result<(Handle, Point)>

Spawns a follower which will follow the given chain and sets its starting point to the given point.