hermes::runtime_extensions::hermes::cron

Module queue

Cron Event Queue implementation.

Structs

CronEventQueue 🔒

The crontab queue task runs in the background.

CronJobDelay 🔒

Cron Job Delay.

Enums

CronJob 🔒

Scheduled Date and Time for sending a cron event.

Functions

new_waiting_task 🔒

Create a new thread that will sleep for duration nanoseconds