Struct hermes::runtime_extensions::hermes::cron::state::InternalState

struct InternalState {
    cron_queue: CronEventQueue,
}

Internal State.

Fields
cron_queue: CronEventQueue

The send events to the crontab queue.

Implementations

impl InternalState

fn new(sender: Option<Sender<CronJob>>) -> Self

Create a new InternalState.

fn add_crontab( &self, app_name: &str, entry: CronTagged, retrigger: bool ) -> bool

Add a new crontab entry.

Allows for management of scheduled cron events queue.

Cron events will be delivered to the on-cron event handler.

Parameters

app_name: AppName. The name of the application that owns the crontab.

entry: CronTagged. The crontab entry to add.

retrigger: bool. If true, the event will re-trigger every time the crontab entry matches until cancelled.

Returns

true: Crontab added successfully.

false: Crontab failed to be added.

fn delay_crontab( &self, app_name: &str, duration: u64, tag: String ) -> Result<bool>

Schedule a single cron event after a fixed delay.

Allows for easy timed wait events to be delivered without requiring datetime calculations or formatting cron entries.

Parameters

app_name: AppName. The name of the application that owns the crontab.

duration: Instant. How many nanoseconds to delay. The delay will be AT LEAST this long.

tag: CronEventTag. A tag which will accompany the triggered event.

Returns

Ok(true): Crontab added successfully.

Ok(false): Crontab failed to be added.

Err: Returns error if the duration is invalid for generating a crontab entry.

fn ls_crontabs( &self, app_name: &str, tag: Option<String> ) -> Vec<(CronTagged, bool)>

List crontabs for an application.

Allows for management of scheduled cron events queue. If tag is none then all crontabs are listed. Otherwise, only the crontabs with the specified tag are listed.

Parameters

tag: Optional, the tag to limit the list to. If none then all crons listed.

Returns

A list of tuples containing the scheduled crontabs and their tags, along with the current retrigger flag. Vec<(CronEventTag, bool)> The list is sorted from most crontab that will trigger soonest to latest. Crontabs are only listed once, in the case where a crontab may be scheduled may times before a later one.

fn rm_crontab(&self, app_name: &str, entry: CronTagged) -> bool

Remove the requested crontab.

Allows for management of scheduled cron events.

Parameters

when: The crontab entry to add. Standard crontab format.

tag: A tag which will accompany the triggered event.

Returns

true: The requested crontab was deleted and will not trigger.

false: The requested crontab does not exist.