hermes::runtime_extensions::hermes::crypto

Module bip32_ed25519

Implementation of Bip32-Ed25519.

Functions

array_u8_32_to_tuple 🔒

Convert a 32 bytes array to a tuple of u64 values.

array_u8_64_to_tuple 🔒

Convert a 64 bytes array to a tuple of u64 values.

b512_u64_tuple_to_u8_array 🔒

Convert a tuple of u64 values to a 64 bytes array.

check_signature 🔒

Check the signature on the given data.

derive_new_private_key 🔒

Derive a new extended private key from the given extended private key.

get_public_key 🔒

Get public key from the given extended private key.

sign_data 🔒

Sign data with the given extended private key.