Module hermes::runtime_extensions::hermes::crypto::bip39

Implementation of BIP39.

Functions

bits_to_bytes 🔒

Turns bit of array to bytes of arrays

byte_to_bit 🔒

Convert bytes entropy to bits.

decimal_to_binary_array 🔒

Turns decimal into binary array of length 11.

generate_entropy 🔒

Generate entropy bytes and return the value.

generate_new_mnemonic 🔒

Generate a new mnemonic. Number of bytes entropy required is calculated by word_count / 3 * 4. eg. 24 words will have 32 bytes entropy or 256 bits entropy. Number of checksum required is calculated by word_count / 3. eg. 24 words will have 8 checksum bits.

get_check_sum_bits 🔒

Generate checksum bits from entropy bits.

get_mnemonic_from_indices 🔒

Get the mnemonic from the BIP39 dictionary using word indices.

get_prefix_index_bits 🔒

Get the index bits of the prefix words from a BIP39 dictionary.

get_word_indices 🔒

Get the word indices from the entropy bits.

is_invalid_word_count 🔒

Check if the word count is invalid. Valid word count is a multiple of 3 and in the range of 12 - 24. Returns true if the word count is invalid, otherwise false.

mnemonic_to_xprv 🔒

Generate a new extended private key (XPrv) from a mnemonic and passphrase.

string_to_language 🔒

Convert string to BIP39 language.