hermes::runtime_extensions::hermes::crypto::bip39

Function generate_new_mnemonic

pub(crate) fn generate_new_mnemonic(
    word_count: usize,
    prefix: Vec<String>,
    language: Option<String>,
) -> Result<Vec<String>, Errno>

Generate a new mnemonic. Number of bytes entropy required is calculated by word_count / 3 * 4. eg. 24 words will have 32 bytes entropy or 256 bits entropy. Number of checksum required is calculated by word_count / 3. eg. 24 words will have 8 checksum bits.

Arguments

word_count: The number of words in the mnemonic. Must be a multiple of 3 and in the range of 12 - 24.

prefix: A vector of strings representing the prefix, empty if no prefix.

language: The language of the mnemonic.

Returns

Returns the mnemonic of type String as a Result. If the conversion is successful, it returns Ok with the mnemonic. If there is an error during the computation, it returns Err with an Errno.

Errors

InvalidMnemonicLength: If the word count is not a multiple of 3 or not in the range of 12 - 24.

PrefixTooLong: If the prefix is longer than the maximum allowed length, max is 3.

WordNotFound: If a word in the mnemonic is not found in the word list.