Module hermes::runtime_extensions::hermes::crypto::state

Crypto state

Structs

ResourceHolder 🔒

Resource holder to hold the resources for XPrv.

WrappedXPrv 🔒

Wrapper for XPrv to implement Hash used in DashMap

Statics

CRYPTO_INTERNAL_STATE 🔒

Global state to hold the resources.

Functions

add_resource 🔒

Add the resource of XPrv to the state if possible. Return the id if successful.

delete_resource 🔒

Delete the resource from the state using id if possible.

get_resource 🔒

Get the resource from the state using id if possible.

get_state 🔒

Get the state.

set_state 🔒

Set the state according to the app context.

Type Aliases

State 🔒

Map of app name to resource holder