Function hermes::runtime_extensions::hermes::hash::blake2b::blake2b_impl

pub(crate) fn blake2b_impl(
    buf: &Vec<u8>,
    outlen: Option<u8>,
) -> Result<Vec<u8>, Errno>

Implementation of the Blake2b hash function.

Arguments

buf: A reference to the byte string for which the hash is to be computed.

outlen: Optional output length in bytes. If not specified, it defaults to 64 bytes.

Returns

Returns the Blake2b hash of the byte string as a Result. If the hash computation is successful, it returns Ok with the hash value. If there is an error during the computation, it returns Err with an Errno.

Errors

InvalidDigestByteLength: If outlen is 0.

HashTooBig: If outlen is greater than 64.