hermes::runtime_extensions::hermes::hash::blake2b

Function blake2bmac_impl

pub(crate) fn blake2bmac_impl(
    buf: &Vec<u8>,
    outlen: Option<u8>,
    key: &Vec<u8>,
    salt: Option<Vec<u8>>,
    personal: Option<Vec<u8>>,
) -> Result<Vec<u8>, Errno>

Implementation of the Blake2b Message Authentication Code.

Arguments

buf: A reference to the byte string for which the blake2bMac is to be computed.

outlen: Optional output length in bytes. If not specified, it defaults to 64 bytes.

key: A reference to the byte string used as the key for computing the blake2bMac.

salt: Optional salt value. If not specified, it defaults to an empty byte string.

personal: Optional personalization string. If not specified, it defaults to an empty byte string.

Returns

Returns the Blake2bMac hash of the byte string as a Result. If the hash computation is successful, it returns Ok with the hash value. If there is an error during the computation, it returns Err with an Errno.

Errors

KeyTooBig: If the length of the key exceeds the specified output length.

HashTooBig: If the specified output length exceeds 64 bytes.

SaltTooBig: If the length of the salt exceeds 16 bytes.

PersonalTooBig: If the length of the personalization string exceeds 16 bytes.