hermes::runtime_extensions::hermes::http_request

Module tokio_runtime_task

Tokio runtime task module for handling HTTP requests.

Structs

TokioTaskHandle

Tokio runtime task handle for sending HTTP requests.

Enums

Command

Represents a command that can be sent to the Tokio runtime task.

Connection 🔒

Represents a connection to an HTTP or HTTPS server.

Constants

HTTP 🔒

HTTP scheme.

HTTPS 🔒

HTTPS scheme.

Statics

INIT_RUSTLS_CRYPTO 🔒

Initializes the Rustls crypto provider for TLS connections.

Functions

executor 🔒

Tokio runtime task for handling HTTP requests.

handle_command 🔒

Handles a command sent to the Tokio runtime task.

init_rustls_connector 🔒

Initializes the Rustls TLS connector with the default root certificates.

read_to_end_ignoring_unexpected_eof 🔒

Reads all bytes from the reader until EOF, ignoring UnexpectedEof errors.

send_http_request 🔒

Sends an HTTP request and puts the response event on the queue.

spawn

Spawns a Tokio runtime task for handling HTTP requests.

Type Aliases

CommandReceiver 🔒

Represents a receiver for commands sent to the Tokio runtime task.

CommandSender 🔒

Represents a sender for commands sent to the Tokio runtime task.