hermes::runtime_extensions::hermes::integration_test

Module event

Integration test runtime extension event handler implementation for test purpose only.

Structs

OnBenchEvent

On bench event

OnTestEvent

On test event

Enums

EventType

Represents different types of events.

Statics

BENCH_RESULT_QUEUE 🔒

Storing results from calling a bench event.

TEST_RESULT_QUEUE 🔒

Storing results from calling a test event.

Functions

execute_event

Executes an event from a module and returns a testing result.