hermes::runtime_extensions::init

Module trait_app

Hermes runtime extensions key traits (Application Pre Initialization).

Runtime extensions must implement these traits if they require code to execute for resource management or other purposes at various phases of the node and application life cycle.

Structs

RTE_INIT_APP_REGISTRY__TraitReg 🔒

RteApp 🔒

How all Application Init is called.

Statics

IS_RTE_APP_INITIALIZED 🔒

Is the RTE Initialized

IS_RTE_APP_LOCK 🔒

Locks the Initialized DashSet to stop race conditions, but only for specific apps. TODO (SJ): Replace with a true key locked map. https://github.com/input-output-hk/catalyst-internal-docs/issues/39#issuecomment-3240936500

RTE_INIT_APP_REGISTRY 🔒

RTE_INIT_APP_REGISTRY__Build 🔒

RTE_INIT_APP_REGISTRY__STORAGE 🔒

Traits

RteInitApp 🔒

Runtime Extension needs Initialization before an Application is loaded by the Node.