hermes::runtime_extensions::init::trait_app

Trait RteInitApp

pub(crate) trait RteInitApp {
    // Provided methods
    fn init(
        self: Box<Self>,
        name: &ApplicationName,
    ) -> Result<(), RuntimeExtensionErrors> { ... }
    fn priority(&self, init: bool) -> i32 { ... }
    fn fini(
        self: Box<Self>,
        name: &ApplicationName,
    ) -> Result<(), RuntimeExtensionErrors> { ... }
}

Runtime Extension needs Initialization before an Application is loaded by the Node.

MUST be used with:

#[traitreg::register]
impl RteInitApp for MyRte {
  // implementation goes here
}

Provided Methods

fn init( self: Box<Self>, name: &ApplicationName, ) -> Result<(), RuntimeExtensionErrors>

Initialize the Runtime extension at App startup (before any modules are initialized). If it errors, the node does not crash, but the App will not load.

Note: self is not required to be used by the implemented function. It is required because of the dynamic initialization logic.

If self is used, it is consumed. This allows the underlying state to be stored in a global static, or the like.

fn priority(&self, init: bool) -> i32

Returns the priority to be used when calling init and fini.

IF the priority is the same, the order of initialization is non-deterministic. Higher priorities run before lower ones.

Two values are returned

0 is the default priority, and should be used for most runtime extensions that do not care about the order of initialization.

fn fini( self: Box<Self>, name: &ApplicationName, ) -> Result<(), RuntimeExtensionErrors>

Finalize the Runtime extension when the application is fully stopped. Can Error, but will not cause the Node to terminate.

Note: self is not required to be used by the implemented function. It is required because of the dynamic initialization logic.

If self is used, it is consumed. This allows the underlying state to be stored in a global static, or the like.