hermes::runtime_extensions::init::trait_runtime

Trait RteInitRuntime

pub(crate) trait RteInitRuntime {
    // Provided methods
    fn init(self: Box<Self>) -> Result<(), RuntimeExtensionErrors> { ... }
    fn priority(&self, init: bool) -> i32 { ... }
    fn fini(self: Box<Self>) -> Result<(), RuntimeExtensionErrors> { ... }
}

Runtime Extension needs Initialization at Node Startup

MUST be used with:

#[traitreg::register(new)]
impl RteInitRuntime for MyRte {
  // implementation goes here
}

Provided Methods

fn init(self: Box<Self>) -> Result<(), RuntimeExtensionErrors>

Initialize the Runtime extension at node startup. This SHOULD be infallible, but if it does fail, then the node must present the errors reported to the user in a nice and actionable way BEFORE it exits.

Note: self is not required to be used by the implemented function. It is required because of the dynamic initialization logic.

If self is used, it is consumed. This allows the underlying state to be stored in a global static, or the like.

fn priority(&self, init: bool) -> i32

Returns the priority to be used when calling init and fini.

IF the priority is the same, the order of initialization is non-deterministic. Higher priorities run before lower ones.

Two values are returned

0 is the default priority, and should be used for most runtime extensions that do not care about the order of initialization.

fn fini(self: Box<Self>) -> Result<(), RuntimeExtensionErrors>

Finalize the Runtime extension at node startup. Is infallible, because runtime extensions should not fail to initialize.

Note: self is not required to be used by the implemented function. It is required because of the dynamic initialization logic.

If self is used, it is consumed. This allows the underlying state to be stored in a global static, or the like.