hermes::runtime_extensions::utils::mul_add

Trait SaturatingMulAdd

pub trait SaturatingMulAdd<U, V> {
    // Required method
    fn mul_add(&mut self, y: U, z: V);
}

A trait providing a saturating multiply-accumulate:

self + (y * z) with saturating arithmetic, updating Self.

self is the Accumulator, and can be any integer but will saturate at max i128 if its a u128 or larger. Otherwise it will saturate within the bounds of its type.

y and z can be any integer type.

Works across any combination of integer types (signed or unsigned).

Required Methods

fn mul_add(&mut self, y: U, z: V)

Multiply and Accumulate Integers.