Module hermes::runtime_extensions::wasi::io::streams::state

Streams state.

Statics

INPUT_STREAMS_STATE ğŸ”’

Global state to hold the input streams resources.

OUTPUT_STREAMS_STATE ğŸ”’

Global state to hold the input streams resources.

Traits

InputStreamTrait ğŸ”’

Helper super trait for InputStream which wraps a std::io::Read and std::io::Seek.

OutputStreamTrait ğŸ”’

Helper super trait for OutputStream which wraps a std::io::Write and std::io::Seek.

Functions

get_input_streams_state ğŸ”’

Get the input streams state.

get_output_streams_state ğŸ”’

Get the output streams state.

Type Aliases

InputStreams ğŸ”’

Map of app name to input streams resource holder.

OutputStreams ğŸ”’

Map of app name to output streams resource holder.