hermes::vfs

Module permission

Permissions state management of the Hermes virtual file system.

Structs

PermissionNode 🔒

PermissionsTree node type.

PermissionsState 🔒

VFS permissions state stored in the radix tree structure, where the each node relates to a single path element with the permission level. PermissionLevel::ReadAndWrite is a default permission level so if permission was not defined for the path explicitly PermissionsState will return the PermissionLevel::ReadAndWrite permission level for the asked path.

Enums

PermissionLevel 🔒

Permission level type.

Type Aliases

PermissionNodeRef 🔒

Convenient type of the referenced PermissionNode.