hermes::wasm

Module module

WASM module implementation. Wrapper over the wasmtime::Module, wasmtime::Instance etc. structs which define a WASM module abstraction with capability to interact with it.

All implementation based on wasmtime crate dependency.

Structs

BadWASMModuleError 🔒

Bad WASM module error

Module

Structure defines an abstraction over the WASM module It instantiates the module with the provided context data, links all provided imports to the module instance, handles an internal state of the WASM module.

ModuleId 🔒

Module id type

ModuleInstance

Structure defines an abstraction over the WASM module instance. It holds the state of the WASM module along with its context data. It is used to interact with the WASM module.