Struct hermes::wasm::module::Module

pub struct Module {
    pre_instance: InstancePre<HermesRuntimeContext>,
    engine: Engine,
    id: ModuleId,
    exc_counter: AtomicU32,
}

Structure defines an abstraction over the WASM module It instantiates the module with the provided context data, links all provided imports to the module instance, handles an internal state of the WASM module.

The primary goal for it is to make a WASM state immutable along WASM module execution. It means that Module::call_func execution does not have as side effect for the WASM module’s state, it becomes unchanged.

Fields
pre_instance: InstancePre<HermesRuntimeContext>

wasmtime::InstancePre entity

A reason why it is used a wasmtime::InstancePre instead of wasmtime::Instance partially described in this RFC. It separates and optimizes the linkage of the imports to the WASM runtime from the module actual initialization process.

engine: Engine

Engine entity

id: ModuleId

Module id

exc_counter: AtomicU32

Module’s execution counter

Implementations

impl Module

pub fn new(module_bytes: &[u8]) -> Result<Self>

Instantiate WASM module

Errors

BadWASMModuleError

BadEngineConfigError

pub(crate) fn id(&self) -> &ModuleId

Get the module id

pub(crate) fn exec_counter(&self) -> u32

Get the module’s execution counter

pub(crate) fn execute_event( &self, event: &dyn HermesEventPayload, state: HermesRuntimeContext ) -> Result<()>

Executes a Hermes event by calling some WASM function. This function abstraction over actual execution of the WASM function, actual definition is inside HermesEventPayload trait implementation.

For each call creates a brand new wasmtime::Store instance, which means that is has an initial state, based on the provided context for each call.

Errors:

BadWASMModuleError