hermes::wasm

Module patcher

Patcher for WASM files. It injects functions to get memory size and read raw memory bytes into the WASM file.

Structs

Patcher 🔒

Patcher for WASM files.

WasmInternals 🔒

Holds the extracted core module and component part of a WASM.

WatMatch 🔒

Represents a single match of a WAT element.

Enums

WatElementMatcher 🔒

A matcher for WAT elements, either by exact string or by regex.

Constants

COMPONENT_CORE_FUNC_REGEX 🔒

Regex to detect the aliases of core functions in the component part.

COMPONENT_EXPORT_FUNC_REGEX 🔒

Regex to detect the function export definitions in the component part.

COMPONENT_FUNC_REGEX 🔒

Regex to detect the function definitions in the component part.

COMPONENT_INJECTIONS 🔒

A template for the injected types, functions and exports in the component part.

CORE_FUNC_REGEX 🔒

Regex to detect the function definitions in the core module.

CORE_INJECTED_EXPORTS 🔒

A template for the injected exports in the core module.

CORE_INJECTED_FUNCTIONS 🔒

A template for the injected functions in the core module.

CORE_INJECTED_TYPES 🔒

A template for the injected types in the core module.

CORE_MODULE_MARKER 🔒

A string that marks the beginning of a core module.

MAGIC 🔒

Magic string to avoid name collisions with existing functions.