Crate hermes

The Hermes Node.

Modules

app 🔒

Hermes app implementation.

event 🔒

Hermes event’s primitives.

logger 🔒

Setup for logging for the service.

reactor 🔒

Hermes Reactor implementation.

runtime_context 🔒

Hermes runtime context implementation.

runtime_extensions 🔒

Hermes runtime extensions

wasm 🔒

WASM related structures and functions which are specific for the Hermes use case. All implementation based on wasmtime crate dependency.

Constants

DEFAULT_ENV_LOG_LEVEL 🔒

The default value for the log level when not specified.

ENV_LOG_LEVEL 🔒

A parameter identifier specifying the log level.

Functions

build_info 🔒

main 🔒