Crate hermes

The Hermes Node.

Modules

app 🔒

Hermes app implementation.

cli 🔒

CLI interpreter for the service

errors 🔒

Errors module.

event 🔒

Hermes event’s primitives.

hdf5 🔒

Module provides different objects, abstractions for working with HDF5 packages.

ipfs 🔒

Hermes IPFS service.

logger 🔒

Setup for logging for the service.

packaging 🔒

Hermes packaging.

pool 🔒

Global thread pool for parallel WASM module execution.

reactor 🔒

Hermes Reactor implementation.

runtime_context 🔒

Hermes runtime context implementation.

runtime_extensions 🔒

Hermes runtime extensions

utils 🔒

Generally used utility functions.

vfs 🔒

Hermes virtual file system.

wasm 🔒

WASM related structures and functions which are specific for the Hermes use case. All implementation based on wasmtime crate dependency.

Macros

add_rte_error

Adds a RuntimeExtensionError to a RuntimeExtensionErrors container.

run_init_fini

Helper to make running init and fini for a Trait registry easy and consistent.

Functions

main 🔒