Module hermes::event::queue

Hermes event queue implementation.

Structs

HermesEventLoopHandler 🔒

Hermes event queue execution loop thread handler

HermesEventQueue 🔒

Hermes event queue. It is a singleton struct.

Enums

Error 🔒

Hermes event queue error

Statics

EVENT_QUEUE_INSTANCE 🔒

Singleton instance of the Hermes event queue.

Functions

event_dispatch 🔒

Execute a hermes event on the provided module and all necessary info.

event_execution_loop 🔒

Executes Hermes events from the provided receiver .

init 🔒

Creates a new instance of the HermesEventQueue. Runs an event loop thread.

send 🔒

Add event into the event queue

targeted_event_execution 🔒

Executes provided Hermes event filtering by target app and target module.