Trait hermes::event::HermesEventPayload

pub(crate) trait HermesEventPayload: Send + Sync + 'static {
    // Required methods
    fn event_name(&self) -> &str;
    fn execute(&self, module: &mut ModuleInstance) -> Result<()>;
}

A trait for defining the behavior of a Hermes event.

Required Methods

fn event_name(&self) -> &str

Returns the name of the event associated with the payload.

fn execute(&self, module: &mut ModuleInstance) -> Result<()>

Executes the behavior associated with the payload, using the provided executor.

Arguments

executor - The executor to use for executing the payload’s behavior.

Returns

An anyhow::Result indicating the success or failure of the payload execution.