Module hermes::packaging::app

Hermes application package.

Modules

author_payload 🔒

A signature payload object for author.cose Hermes application package. Defined at https://input-output-hk.github.io/hermes/architecture/08_concepts/hermes_packaging_requirements/app_signatures/#wasm-component-module-signatures.

manifest 🔒

Hermes application package manifest.json struct.

Structs

ApplicationPackage 🔒

Hermes application package.

Functions

validate_and_write_from_manifest 🔒

Validate and write all content of the Manifest to the provided package.

validate_and_write_icon 🔒

Validate icon.svg file and write it to the package to the provided dir path.

validate_and_write_metadata 🔒

Validate metadata.json file and write it to the package to the provided dir path. Also updates Metadata object by setting build_date and name properties.

validate_and_write_module 🔒

Validate WASM module package and write it to the package to the provided dir path.

write_share_dir 🔒

Write share dir to the package to the provided dir path.

write_www_dir 🔒

Write www dir to the package to the provided dir path to the provided dir path.