Struct hermes::packaging::app::ApplicationPackage

pub(crate) struct ApplicationPackage(Package);

Hermes application package.

Tuple Fields
0: Package
Implementations

impl ApplicationPackage

const AUTHOR_COSE_FILE: &'static str = "author.cose"

Application package signature file path.

const FILE_EXTENSION: &'static str = "happ"

Hermes application package file extension.

const ICON_FILE: &'static str = "icon.svg"

Hermes application package icon file path.

const METADATA_FILE: &'static str = "metadata.json"

Hermes application package metadata file path.

const MODULES_DIR: &'static str = "lib"

Application WASM modules directory path.

const SHARE_DIR: &'static str = "srv/share"

Application package share directory path.

const USR_DIR: &'static str = "usr"

Application shareable directory path.

const WWW_DIR: &'static str = "srv/www"

Application package www directory path.

pub(crate) fn build_from_manifest<P: AsRef<Path>>( manifest: &Manifest, output_path: P, package_name: Option<&str>, build_date: DateTime<Utc>, ) -> Result<Self>

Create a new Hermes application package package from a manifest file.

pub(crate) fn from_file<P: AsRef<Path>>(path: P) -> Result<Self>

Open an existing application package.

pub(crate) fn validate(&self, untrusted: bool) -> Result<()>

Validate package with its signature and other contents. If untrusted flag is true the signature will not be verified.

fn verify_author_sign(&self) -> Result<()>

Verify author package signature.

pub(crate) fn author_sign( &self, private_key: &PrivateKey, certificate: &Certificate, ) -> Result<()>

Sign the package as an author and store signature inside it. If signature already exists it will be extended with a new signature.

fn get_author_signature_payload(&self) -> Result<SignaturePayload>

Build and return author_payload::SignaturePayload.

pub(crate) fn get_metadata(&self) -> Result<Metadata<Self>>

Get Metadata object from package.

pub(crate) fn get_author_signature( &self, ) -> Result<Option<Signature<SignaturePayload>>>

Get author Signature object from package.

pub(crate) fn get_modules(&self) -> Result<Vec<(String, WasmModulePackage)>>

Get Vec<WasmModulePackage> from package.

pub(crate) fn get_www(&self) -> Option<Dir>

Get www dir from package if present.

pub(crate) fn get_share(&self) -> Option<Dir>

Get share dir from package if present.