hermes::packaging

Module package

Implementation of the generalized Hermes package struct as a HDF5 package.

Structs

Package 🔒

Hermes package object. Wrapper over HDF5 file object.

Constants

BUFFER_SIZE 🔒

Buffer size for hash calculation. 1024 * 1024 = 1MB.

Functions

calculate_dir_hash 🔒

Calculates recursively directory contents hash with the provided hasher including file contents.

calculate_file_hash 🔒

Calculates file hash with the provided hasher.