hermes::runtime_extensions::bindings::hermes::cardano::api

Trait HostTransaction

pub trait HostTransaction {
    // Required methods
    fn get_metadata(
        &mut self,
        self_: Resource<Transaction>,
        label: u64,
    ) -> Result<Option<Vec<u8>>>;
    fn get_txn_hash(
        &mut self,
        self_: Resource<Transaction>,
    ) -> Result<Option<(u64, u64, u64, u64)>>;
    fn raw(&mut self, self_: Resource<Transaction>) -> Result<Option<Vec<u8>>>;
    fn drop(&mut self, rep: Resource<Transaction>) -> Result<()>;
}

Required Methods

fn get_metadata( &mut self, self_: Resource<Transaction>, label: u64, ) -> Result<Option<Vec<u8>>>

Returns the transaction auxiliary metadata in CBOR format.

Parameters

label: A metadata label used as a key to get the associated metadata.

Returns

option<cbor> : The CBOR format of the metadata, None if the label requested is not present.

fn get_txn_hash( &mut self, self_: Resource<Transaction>, ) -> Result<Option<(u64, u64, u64, u64)>>

Returns the transaction hash.

Returns

option<txn-hash> : Cardano transaction hash - Blake2b-256, None if cannot retrieve the transaction hash.

fn raw(&mut self, self_: Resource<Transaction>) -> Result<Option<Vec<u8>>>

Returns the raw CBOR representation of the transaction.

Returns

option<cbor> : The CBOR format of the transaction, None if cannot retrieve the raw transaction.

fn drop(&mut self, rep: Resource<Transaction>) -> Result<()>