Trait hermes::runtime_extensions::bindings::hermes::localtime::api::Host

pub trait Host {
    // Required methods
    fn get_localtime(
        &mut self,
        when: Option<Datetime>,
        tz: Option<String>
    ) -> Result<Result<Localtime, Errno>>;
    fn alt_localtime(
        &mut self,
        time: Localtime,
        tz: Option<String>
    ) -> Result<Result<Localtime, Errno>>;
    fn get_datetime(
        &mut self,
        time: Localtime
    ) -> Result<Result<Datetime, Errno>>;
}

Required Methods

fn get_localtime( &mut self, when: Option<Datetime>, tz: Option<String> ) -> Result<Result<Localtime, Errno>>

Get localtime from a datetime or now.

Parameters

when : The datetime we want to convert (Optional, if not set it will convert now). tz : The timezone to use. (Optional, if not set uses the local machines configured local timezone.)

Returns

localtime : the converted time. errno : An error indicating why conversion failed.

fn alt_localtime( &mut self, time: Localtime, tz: Option<String> ) -> Result<Result<Localtime, Errno>>

Get a new localtime from a localtime, by recalculating time for a new timezone.

Parameters

time : The localtime to convert. tz : The timezone to use. (Optional, if not set uses the local machines configured local timezone.)

Returns

localtime : the converted time. errno : An error indicating why conversion failed.

fn get_datetime(&mut self, time: Localtime) -> Result<Result<Datetime, Errno>>

Get a datetime from a localtime.

Parameters

time : The localtime to convert.

Returns

datetime : the converted time. errno : An error indicating why conversion failed.