Trait hermes::runtime_extensions::bindings::hermes::sqlite::api::HostSqlite

pub trait HostSqlite {
    // Required methods
    fn close(&mut self, self_: Resource<Sqlite>) -> Result<Result<(), Errno>>;
    fn status(
        &mut self,
        self_: Resource<Sqlite>,
        opt: StatusOptions,
        reset_flag: bool
    ) -> Result<Result<(i32, i32), Errno>>;
    fn prepare(
        &mut self,
        self_: Resource<Sqlite>,
        sql: String
    ) -> Result<Result<Resource<Statement>, Errno>>;
    fn execute(
        &mut self,
        self_: Resource<Sqlite>,
        sql: String
    ) -> Result<Result<(), Errno>>;
    fn drop(&mut self, rep: Resource<Sqlite>) -> Result<()>;
}

Required Methods

fn close(&mut self, self_: Resource<Sqlite>) -> Result<Result<(), Errno>>

Closes a database connection, destructor for sqlite3.

Ideally, applications should finalize all prepared statements associated with the sqlite3 object prior to attempting to close the object. If the database connection is associated with unfinalized prepared statements, then the function will leave the database connection open and return the busy error code.

If an sqlite3 object is destroyed while a transaction is open, the transaction is automatically rolled back.

fn status( &mut self, self_: Resource<Sqlite>, opt: StatusOptions, reset_flag: bool ) -> Result<Result<(i32, i32), Errno>>

Retrieves runtime status information about a single database connection.

Parameters

opt: An integer constant, taken from the set of status-options, that determines the parameter to interrogate.

reset-flag: If is true, then the highest instantaneous value is reset back down to the current value.

Returns

A tuple of the current value of the requested parameter, and the highest instantaneous value on success, and an error code on failure.

fn prepare( &mut self, self_: Resource<Sqlite>, sql: String ) -> Result<Result<Resource<Statement>, Errno>>

Compiles SQL text into byte-code that will do the work of querying or updating the database.

Parameters

db: Database handle.

sql: SQL statement, UTF-8 encoded.

Returns

A compiled prepared statement that can be executed using sqlite3_step(). If there is an error or the input text contains no SQL (if the input is an empty string or a comment) then an error code is returned.

fn execute( &mut self, self_: Resource<Sqlite>, sql: String ) -> Result<Result<(), Errno>>

Executes an SQL query directly without preparing it into a statement and returns the result.

Parameters

sql: SQL statement, UTF-8 encoded.

fn drop(&mut self, rep: Resource<Sqlite>) -> Result<()>