Trait hermes::runtime_extensions::bindings::wasi::http::outgoing_handler::Host

pub trait Host {
    // Required method
    fn handle(
        &mut self,
        request: Resource<OutgoingRequest>,
        options: Option<Resource<RequestOptions>>,
    ) -> Result<Result<Resource<FutureIncomingResponse>, ErrorCode>>;
}

Required Methods

fn handle( &mut self, request: Resource<OutgoingRequest>, options: Option<Resource<RequestOptions>>, ) -> Result<Result<Resource<FutureIncomingResponse>, ErrorCode>>

This function is invoked with an outgoing HTTP Request, and it returns a resource future-incoming-response which represents an HTTP Response which may arrive in the future.

The options argument accepts optional parameters for the HTTP protocol’s transport layer.

This function may return an error if the outgoing-request is invalid or not allowed to be made. Otherwise, protocol errors are reported through the future-incoming-response.