Trait hermes::runtime_extensions::bindings::wasi::io::streams::HostInputStream

pub trait HostInputStream {
    // Required methods
    fn read(
        &mut self,
        self_: Resource<InputStream>,
        len: u64
    ) -> Result<Result<Vec<u8>, StreamError>>;
    fn blocking_read(
        &mut self,
        self_: Resource<InputStream>,
        len: u64
    ) -> Result<Result<Vec<u8>, StreamError>>;
    fn skip(
        &mut self,
        self_: Resource<InputStream>,
        len: u64
    ) -> Result<Result<u64, StreamError>>;
    fn blocking_skip(
        &mut self,
        self_: Resource<InputStream>,
        len: u64
    ) -> Result<Result<u64, StreamError>>;
    fn drop(&mut self, rep: Resource<InputStream>) -> Result<()>;
}

Required Methods

fn read( &mut self, self_: Resource<InputStream>, len: u64 ) -> Result<Result<Vec<u8>, StreamError>>

Perform a non-blocking read from the stream.

This function returns a list of bytes containing the read data, when successful. The returned list will contain up to len bytes; it may return fewer than requested, but not more. The list is empty when no bytes are available for reading at this time. The pollable given by subscribe will be ready when more bytes are available.

This function fails with a stream-error when the operation encounters an error, giving last-operation-failed, or when the stream is closed, giving closed.

When the caller gives a len of 0, it represents a request to read 0 bytes. If the stream is still open, this call should succeed and return an empty list, or otherwise fail with closed.

The len parameter is a u64, which could represent a list of u8 which is not possible to allocate in wasm32, or not desirable to allocate as as a return value by the callee. The callee may return a list of bytes less than len in size while more bytes are available for reading.

fn blocking_read( &mut self, self_: Resource<InputStream>, len: u64 ) -> Result<Result<Vec<u8>, StreamError>>

Read bytes from a stream, after blocking until at least one byte can be read. Except for blocking, behavior is identical to read.

fn skip( &mut self, self_: Resource<InputStream>, len: u64 ) -> Result<Result<u64, StreamError>>

Skip bytes from a stream. Returns number of bytes skipped.

Behaves identical to read, except instead of returning a list of bytes, returns the number of bytes consumed from the stream.

fn blocking_skip( &mut self, self_: Resource<InputStream>, len: u64 ) -> Result<Result<u64, StreamError>>

Skip bytes from a stream, after blocking until at least one byte can be skipped. Except for blocking behavior, identical to skip.

fn drop(&mut self, rep: Resource<InputStream>) -> Result<()>