Module hermes::runtime_extensions::hermes::cardano

Cardano Blockchain runtime extension implementation.

Modules

chain_follower_task 🔒

A Chain Follower task is responsible for managing a Cardano Chain Follower that is controlled by the Cardano Runtime Extension.

event 🔒

Cardano Blockchain runtime extension event handler implementation.

host 🔒

Cardano Blockchain host implementation for WASM runtime.

tokio_runtime_task 🔒

The Tokio Runtime task is responsible for executing a Tokio runtime inside a OS thread so that Futures can be executed by the Cardano Runtime Extension.

Structs

State 🔒

Cardano Runtime Extension state.

SubscriptionState 🔒

Hermes application module subscription state.

Enums

SubscriptionType 🔒

Available subscription types.

Statics

STATE 🔒

Cardano Runtime Extension internal state.

Functions

new_context 🔒

Advise Runtime Extensions of a new context

read_block 🔒

Reads a block from a Cardano network.

subscribe 🔒

Subscribes a module or resumes the generation of subscribed events for a module.

unsubscribe 🔒

Unsubscribes a module or stops the generation of subscribed events for a module.

Type Aliases

ModuleStateKey 🔒

Triple representing the key of the subscription state map.

Result 🔒

Cardano Runtime Extension internal result type.