Module hermes::runtime_extensions::hermes::cardano::chain_follower_task

A Chain Follower task is responsible for managing a Cardano Chain Follower that is controlled by the Cardano Runtime Extension.

Structs

EventSubscriptions 🔒

Holds flags specifying which event subscriptions are active.

Handle

Handle used to communicate with a chain follower executor task.

Enums

Command 🔒

Chain follower executor commands.

Functions

build_and_send_block_event 🔒

Builds a super::event::OnCardanoBlockEvent from the block data and sends it to the given module through the Event Queue.

build_and_send_rollback_event 🔒

Builds a super::event::OnCardanoRollback from the block data and sends it to the given module through the Event Queue.

build_and_send_txns_event 🔒

Builds super::event::OnCardanoTxnEvent for every transaction on the block data and sends them to the given module through the Event Queue.

executor 🔒

Runs a Cardano chain follower that generates events for the given application module and is connected to the given chain.

get_event_subscriptions 🔒

Gets the event subscription flags for a given module.

process_block_chain_update 🔒

Processes a block chain update.

process_chain_update 🔒

Processes a single chain update.

process_command 🔒

Processes a chain follower task command.

process_rollback_chain_update 🔒

Processes a rollback chain update.

spawn

Spawns a new Chain Follower task in the current Tokio runtime.

update_current_slot 🔒

Updates the module’s state with the current slot the follower is at.

Type Aliases

CommandReceiver 🔒

Chain follower handle command channel receiver.

CommandSender 🔒

Chain follower handle command channel sender.