Module hermes::runtime_extensions::hermes::cardano::tokio_runtime_task

The Tokio Runtime task is responsible for executing a Tokio runtime inside a OS thread so that Futures can be executed by the Cardano Runtime Extension.

Structs

Handle

Handle used for communicating with the Tokio runtime background thread.

Enums

Command 🔒

Command data that can be send to the Tokio runtime background thread.

Functions

executor 🔒

Runs the Cardano Runtime Extension Tokio runtime.

follower_connect_address 🔒

Returns the peer address used to connect to each Cardano network.

read_block 🔒

Reads a block from the given chain at the given point.

spawn

Spawns a OS thread running the Tokio runtime task.

spawn_follower 🔒

Spawns a follower which will follow the given chain and sets its starting point to the given point.

Type Aliases

CommandReceiver 🔒

Tokio runtime handle command channel receiver type.

CommandSender 🔒

Tokio runtime handle command channel sender type.