Function hermes::runtime_extensions::hermes::crypto::bip32_ed25519::derive_new_private_key

pub(crate) fn derive_new_private_key(
    xprivate_key: XPrv,
    path: &str,
) -> Result<XPrv, Errno>

Derive a new extended private key from the given extended private key.

V2 derivation scheme is used as it is mention in SLIP-0023.

More information about child key derivation can be found in BIP32-Ed25519.

Arguments

xprivate_key: An extended private key of type XPrv.

path: Derivation path. eg. m/0/2’/3 where ’ represents hardened derivation.

Returns

Returns the XPrv extended private key as a Result. If the derivation path is successful, it returns Ok with the extended private key (XPrv).

Errors

Returns an InvalidDerivationalPath if the derivation path is invalid.