Module hermes::runtime_extensions::hermes::http_gateway::routing

Gateway routing logic

Structs

Hostname 🔒

hostname (node name)

Constants

EVENT_TIMEOUT 🔒

Attempts to wait for a value on this receiver, returning an error if the corresponding channel has hung up, or if it waits more than timeout of arbitrary 1 second

VALID_PATH 🔒

Check path is valid for static files

WEBASM_ROUTE 🔒

Everything that hits /api routes to Webasm Component Modules

Functions

compose_http_event 🔒

Compose http event and send to global queue, await queue response and relay back to waiting receiver channel for HTTP response

error_response 🔒

HTTP error response generator

host_resolver 🔒

Extractor that resolves the hostname of the request. Hostname is resolved through the Host header

is_valid_path 🔒

Check if valid path to static files.

not_found 🔒

HTTP not found response generator

route_to_hermes 🔒

Route single request to hermes backend

router 🔒

Routing by hostname is a mechanism for isolating API services by giving each API its own hostname; for example, service-a.api.example.com or service-a.example.com.

serve_static_data 🔒

Serves static data with 1:1 mapping