Module hermes::runtime_extensions::hermes::sqlite::state

Internal state implementation for the SQLite module.

Statics

SQLITE_DB_STATE 🔒

Global state to hold SQLite db resources.

SQLITE_STATEMENT_STATE 🔒

Global state to hold SQLite statement resources.

Functions

get_db_state 🔒

Get the global state of SQLite db resources.

get_statement_state 🔒

Get the global state of SQLite statement resources.

Type Aliases

DbState 🔒

Map of app name to db resource holder

ObjectPointer 🔒

The object pointer used specifically with C objects like sqlite3 or sqlite3_stmt.

StatementState 🔒

Map of app name to db statement resource holder